Edge overlap heatmaps for a set of AntibodyForests sequence similarity networks or minimum spanning trees.

Similar to the AntibodyForests_node_transitions function, will calculate the incidence of features across undirected edges. In this case, each edge will be considered a unique species - with incidence counts across each unique feature value if a specific edge is connected to a node with that feature. Overlap metrics are then calculated for this edge-feature incidence matrix.

AntibodyForests_overlap(trees, group.by, method)

Arguments

trees

nested list of AntibodyForests objects, as obtained from the AntibodyForests function.

group.by

vector of strings - node features to group the edges by (counts edge incidence across the unique feature values for the specified node feature).

method

string - overlap calculator: 'overlap' for unique/public edge counts across the feature values, 'jaccard' to calculate the Jaccard index.

Value

Edge overlap heatmaps for the specific overlap metric/method.

Examples

if (FALSE) {
AntibodyForests_overlap(trees, group.by = 'seurat_clusters', method = 'jaccard')
}