Coexpression of selected genes

Returns eiter a plot or numeric data of coexpression levels of selected genes.Coexpression % is calculated as the quotient of double positive cells (counts \> 0) and the sum of total cells positive for either genes.

GEX_coexpression_coefficient(GEX, genes, subsample.n, plot.dotmap)

Arguments

GEX

GEX seurat object generated with VDJ_GEX_matrix (VDJ_GEX_matrix.output\[\[2\]\])

genes

Character vector. At least 2 genes present in rownames(GEX). Use "all" to include all genes. The number of comparisons to make is the length(genes)! (factorial). More than 100 genes are not recommended.

subsample.n

Interger. Number of cells to subsample. If set to 100, 100 cells will be randomly sampled for the calculation

plot.dotmap

Boolean. Whether to return a plot

Value

Returns a dataframe if pot.dotmap == F or a ggplot if plot.dotmap == T detailing the coexpression levels of selected genes within the given cell population

Examples

#To return a dataframe with coefficients
#GEX_coexpression_coefficient(GEX = VDJ_GEX_matrix.output[[2]]
#, genes = c("CD19", "EBF1","SDC1"), subsample.n = "none", plot.dotmap = FALSE)

#To return a dotplot detailing coexpression and overall expression
GEX_coexpression_coefficient(GEX = Platypus::small_vgm[[2]]
, genes = c("CD19", "CD83"), subsample.n = "none", plot.dotmap = FALSE)
#> Calculating coexpression for 2 genes with 21 cells
#>   col.1 col.2 gene.1 gene.2 coex.coef perc.single.positive
#> 1     1     2   CD19   CD83  41.93548             41.93548