This function generates new db lookup and user lookup tables from the PlatypusDB. This is normally done within the Platypus_DB_upload function. In case of manual changes to files, incomplete uploads or other changes after the last use of PlatypusDB_upload, this function may be used to generate up-to-date lookup tables. To use, you will need writing permission from this R session

This function generates new db lookup and user lookup tables from the PlatypusDB. This is normally done within the Platypus_DB_upload function. In case of manual changes to files, incomplete uploads or other changes after the last use of PlatypusDB_upload, this function may be used to generate up-to-date lookup tables. To use, you will need writing permission from this R session

PlatypusDB_refresh_lookup(bucket.name)

Arguments

bucket.name
Character. Name of bucket containing data in. Defaults to "platypusdb_cellrangerv6_data" (This is a feature to support future changes to the database)

Value

Writes a .csv file to the current working directory, which contains the lookup table for the website. The lookup for the R interface will be uploaded to PlatypusDB and returned for control

Examples

if (FALSE) {
new_table <- PlatypusDB_refresh_lookup(bucket.name = "bucket with database data")
}