Pseudotime analysis for scRNA and repertoire sequencing datasets

Pseudotime analysis for scRNA and repertoire sequencing datasets

VDJ_GEX_clonotyme(
  method,
  version,
  top.N.clonotypes,
  vdj.gex.matrix.output,
  vdj.analyze.output,
  gex.automate.output,
  exclude.clusters,
  colors,
  show.cells,
  highlight.genes,
  dropest.output.list,
  velocyto.gex.merged,
  velocyto.file.name,
  velocyto.out.dir,
  velocyto.save.rds,
  velocyto.norm.scale.factor,
  velocyto.n.variable.features,
  velocyto.neighbor.dim,
  velocyto.cluster.resolution,
  velocyto.mds.dim,
  velocyto.nCount_spliced,
  velocyto.percent.mt,
  velocyto.normalisation.method,
  velocyto.selection.method,
  velocyto.deltaT,
  velocyto.kCells,
  velocyto.fit.quantile,
  velocyto.kGenes,
  root.selection,
  root.marker,
  ridgeline.separator,
  genes.for.module.score,
  root.nodes,
  color.cells
)

Arguments

method

Pseudotime analylysis method to be used. Possible parameters are monocle3 or velocyto. monocle3 is being used as a default method. For velocyto analysis please run on Cluster and it is only available for UNIX based systems.

version

Platypus version to use "v2" or "v3". version 2 used by default.

top.N.clonotypes

How many clonotypes to show per sample in the Ridgeline plots and on the Velocyto UMAP.

vdj.gex.matrix.output

If Platypus v3 is used, the input to this function has to be the output of the VDJ_GEX_matrix function.

vdj.analyze.output

If Platypus v2 is used, the VDJ_analyze output has to be supplied.

gex.automate.output

If Platypus v2 is used, the GEX_automate output has to be supplied here.

exclude.clusters

Please enter a cluster number if you'd like to exclude a certain cluster from analysis. Cluster will be assigned to different partition in Monocle3 analysis and therefore pseudotime distance will be set to infinity. Cells from this cluster will be deleted from the dataset in the Velocyto analysis.

colors

Vector containing custom colors to be used for highlighting the clonotypes. If left empty, default colors will be assigned.

show.cells

Logical, should cells be shown in the Ridgeline plots. True by default.

highlight.genes

Vector containing gene names. The expressionlevels of these genes along pseudotime will be plotted.

dropest.output.list

List containing the cell.counts.matrices.rds from the Dropest alignment for Velocyto analysis.

velocyto.gex.merged

Logical whether samples should be shown in combined UMAP or sepeartely.

velocyto.file.name

String used as file name when saving the output pdf

velocyto.out.dir

Directory to save the output files. By default the current working directory.

velocyto.save.rds

If RDS objects should be saved as well. Default = F.

velocyto.norm.scale.factor

Parameter for GEX analysis of the cell.count.matrices.

velocyto.n.variable.features

Parameter for GEX analysis of the cell.count.matrices.

velocyto.neighbor.dim

Parameter for GEX analysis of the cell.count.matrices.

velocyto.cluster.resolution

Parameter for GEX analysis of the cell.count.matrices.

velocyto.mds.dim

Parameter for GEX analysis of the cell.count.matrices.

velocyto.nCount_spliced

Cutoff threshold. cells with less spliced gene counts will be omitted. Filtering of bad quality cells.

velocyto.percent.mt

Parameter for GEX analysis of the cell.count.matrices.

velocyto.normalisation.method

Parameter for GEX analysis of the cell.count.matrices.

velocyto.selection.method

Parameter for GEX analysis of the cell.count.matrices.

velocyto.deltaT

Parameter for Velocyto analysis

velocyto.kCells

Parameter for Velocyto analysis

velocyto.fit.quantile

Parameter for Velocyto analysis

velocyto.kGenes

Parameter for Velocyto analysis

root.selection

Character. Method for root selection. Defaults to "manual"

root.marker

Character. Marker to use as Root. Defaults to "SELL"

ridgeline.separator

Character. Variable to group ridgeline plots by. Defaults to "clonotype"

genes.for.module.score

List of vectors of genes. With module scores inferred via Seurat::AddModuleScore(). Default is set to NULL.

root.nodes

Labeled node from trajectory plot to specify root nodes root nodes for pseudotime trajectory.

color.cells

For the module score plot decide how the cells should be coloured (based on e.g. group_id, seurat_clusters etc.). Default = 'seurat_clusters'

Value

If method=monocle3, the function returns a list element: [[1]] UMAP colored by Pseudotime, [[2]] Ridgeline plots showing the density of each of the top.N.clonotypes per cluster along pseudotime., [[3]] Gene expression plots highlighting the gene expression across pseudotime colored by transcriptional cluster, [[4]] Gene expression plots highlighting the gene expression across pseudotime colored by colotype. If method=velocyto, plots and RDS will be saved to velocyto.out.dir.

Examples