Gives stats on number and quality of reads.

Gives stats on number and quality of reads.

VDJ_GEX_stats(
  VDJ.out.directory,
  GEX.out.directory,
  sample.names,
  metrics10x,
  save.csv,
  filename
)

Arguments

VDJ.out.directory
List of paths with each element containing the path to the output of cellranger VDJ runs. This pipeline assumes that the output file names have not been changed from the default 10x settings in the /outs/ folder. This is compatible with B and T cell repertoires (both separately and simultaneously).

GEX.out.directory
OPTIONAL list of paths with each element containing the path to the output of cellranger GEX runs. This pipeline assumes that the output file names have not been changed from the default 10x settings in the /outs/ folder. This is compatible with B and T cell repertoires (both separately and simultaneously).

sample.names
OPTIONAL: an array of the same length as the input VDJ.out.directory list with custom names for each sample. If not provided samples will be numbered by processing order

metrics10x
Whether to append metrics_summary.csv information provided by Cellranger for both VDJ and GEX. Defaults to T

save.csv
Boolean. Defaults to TRUE. Whether to directly save the resuts as a comma delimited .csv file in the current working directory.

filename
Character ending in .csv. Filename to save .csv as.

Value

returns a single matrix where the rows are individual cells and the columns are repertoire features.

Examples

if (FALSE) {
stats <- VDJ_GEX_stats(VDJ.out.directory = VDJ.out.directory.list
,GEX.out.directory = GEX.out.directory.list,sample.names = c(1:4)
,metrics10x = T,save.csv = T ,filename = "stats.csv")
}