Calculate abundances/counts of specific features for a VDJ dataframe

Calculate the absolute counts or proportions of a specific cell-level feature (column in the VDJ/VDJ.GEX.matrix[[1]] object), per an optional specific grouping factor (e.g., clonotype via 'clonotype_id') and an optional sample factor(e.g., 'sample_id'). Outputs either a count dataframe of the specific feature or a ggplot2 barplot.

VDJ_abundances(
  VDJ,
  feature.columns,
  proportions,
  specific.features,
  grouping.column,
  max.groups,
  specific.groups,
  sample.column,
  VDJ.VJ.1chain,
  treat.incomplete.groups,
  treat.incomplete.features,
  combine.features,
  treat.combined.features,
  specific.feature.colors,
  output.format
)

Arguments

VDJ
VDJ or VDJ.GEX.matrix[[1]] object, as obtained from the VDJ_GEX_matrix function in Platypus.

feature.columns
vector of strings, denoting the columns of the VDJ/VDJ.GEX.matrix[[1]] object from which to extract the unique feature values (for which we will calculate the counts or proportions).

proportions
string, 'absolute' will return the absolute counts, 'group.level.proportions' will return the counts divided by the total number or elements/values in the specific groups (group level proportions), 'sample.level.proportions' will return the counts divided by the total number of elements in the sample.

specific.features
vector of specific feature values (or NULL) for which to calculate counts/proportions, from the specified feature.columns parameter (only works if a single feature column is specified in feature.columns).

grouping.column
string or 'none', represents the column from the VDJ/VDJ.GEX.matrix[[1]] object by which to group counting process. This is usually the 'clonotype_id' column to calculate frequencies at the clonotype level. If 'none', no grouping will be done. For example, if feature.columns='VDJ_cgene' and grouping.column='clonotype_id', we will obtain a count dataframe of the frequencies of each isotype per unique clonotype (per sample if sample.column='sample_id').

max.groups
integer or NULL, the maximum number of groups for which to count features. If NULL, it will count for all groups.

specific.groups
vector of strings (or 'none'), if the counting should be done only for specific groups (e.g., count the frequency of isotype only for clonotypes 1 and 2 if feature.columns='VDJ_cgene', grouping.column='clonotype_id' and specific.groups=c('clonotype1', 'clonotype2'))

sample.column
string, represents the sample column if your VDJ/VDJ.GEX.matrix[[1]] object has multiple samples (usually 'sample_id')

VDJ.VJ.1chain
boolean, if T will remove aberrant cells (more than 1 VDJ of VJ chain), if F it will keep them.

treat.incomplete.groups
string, method of dealing with groups which are missing the features in the feature.columns parameter (e.g., a clonotype which does not have any transcriptomic clusters annotations if feature.columns='transcript_cluster').'exclude' - excludes groups with no cells for the specific features, 'unknown' - sets them as unknown

treat.incomplete.features
string, method of dealing with missing feature values (e.g., a clonotype has several NA values for the 'VDJ_cgene' feature.column - cells with NA values). 'unknown' - counted as unknown, 'exclude' - excludes completely, 'max.global' - replaces value by max value of that feature across the repertoire, 'max.group' - replaced by the max feature value inside that group, 'proportional' - iteratively assigns the missing values to the known groups, keeping the same proportions.

combine.features
boolean - if T and we have two columns in feature.columns, will combine the feature values for each cell in the VDJ object, counting them as a single feature when calculating proportions.

treat.combined.features
string, method of dealing with combined features with missing values. 'exclude' will be treated similarly to excluding incomplete feature values (excluding them completely if a single value is missing from the combination), or 'include' and will be treated as a new feature value.

specific.feature.colors
named list of specific colors to be used in the final barplots, for each unique feature value in the VDJ object's feature.columns values. For example, if we have a feature column of binders with unique values=c('yes', 'no'), specific.feature.colors=list('yes'='blue', 'no'='red') will color them accordingly.

output.format
string, either 'plots' to obtain barplots, 'abundance.df' to obtain the count dataframe, or 'abundance.df.list' to obtain a list of count dataframes, for each sample.

Value

Either a count dataframe with the following columns: group(=unique group value, e.g., 'clonotype1' if grouping.column='clonotype_id'), sample, group_frequency, unique_feature_values, feature_value_counts, total_feature_names or a barplot of the counts/proportions per feature, per group.

Examples

VDJ_abundances(VDJ = small_vgm[[1]],
feature.columns='VDJ_cgene',proportions='absolute',
grouping.column='clonotype_id',specific.groups='none',
output.format='abundance.df')
#>            group sample sample_frequency group_frequency unique_feature_values
#> 1  clonotype2053     s1               17               1                IGHG2C
#> 2  clonotype2386     s1               17               1                IGHG2C
#> 3  clonotype2040     s1               17               1                  IGHM
#> 4  clonotype1725     s1               17               1                IGHG2C
#> 5  clonotype1243     s1               17               1                  IGHM
#> 6  clonotype2271     s1               17               1                  IGHD
#> 7  clonotype1364     s1               17               1                  IGHM
#> 8  clonotype2368     s1               17               1                  IGHA
#> 9  clonotype2069     s1               17               1                  IGHA
#> 10 clonotype1514     s1               17               1                  IGHM
#> 11 clonotype1682     s1               17               1                  IGHM
#> 12 clonotype2276     s1               17               1                  IGHM
#> 13 clonotype1451     s1               17               1                  IGHM
#> 14 clonotype2295     s1               17               1                  IGHM
#> 15 clonotype2100     s1               17               1                  IGHM
#> 16 clonotype2287     s1               17               1                  IGHM
#> 17 clonotype1923     s1               17               1                  IGHM
#> 18  clonotype141     s3               11               1                  IGHM
#> 19  clonotype390     s3               11               1                  IGHM
#> 20  clonotype464     s3               11               1                  IGHM
#> 21 clonotype1028     s3               11               1                  IGHD
#> 22  clonotype592     s3               11               1                  IGHM
#> 23  clonotype777     s3               11               1                  IGHM
#> 24  clonotype921     s3               11               1                  IGHM
#> 25  clonotype897     s3               11               1                  IGHM
#> 26 clonotype1056     s3               11               1                 IGHG3
#> 27  clonotype649     s3               11               1                  IGHM
#> 28  clonotype478     s3               11               1                  IGHM
#> 29  clonotype649     s2                5               1                  IGHM
#> 30  clonotype634     s2                5               1                  IGHM
#> 31  clonotype686     s2                5               1                  IGHD
#> 32  clonotype734     s2                5               1                  IGHM
#> 33  clonotype635     s2                5               1                  IGHM
#>    feature_value_counts feature_name
#> 1                     1    VDJ_cgene
#> 2                     1    VDJ_cgene
#> 3                     1    VDJ_cgene
#> 4                     1    VDJ_cgene
#> 5                     1    VDJ_cgene
#> 6                     1    VDJ_cgene
#> 7                     1    VDJ_cgene
#> 8                     1    VDJ_cgene
#> 9                     1    VDJ_cgene
#> 10                    1    VDJ_cgene
#> 11                    1    VDJ_cgene
#> 12                    1    VDJ_cgene
#> 13                    1    VDJ_cgene
#> 14                    1    VDJ_cgene
#> 15                    1    VDJ_cgene
#> 16                    1    VDJ_cgene
#> 17                    1    VDJ_cgene
#> 18                    1    VDJ_cgene
#> 19                    1    VDJ_cgene
#> 20                    1    VDJ_cgene
#> 21                    1    VDJ_cgene
#> 22                    1    VDJ_cgene
#> 23                    1    VDJ_cgene
#> 24                    1    VDJ_cgene
#> 25                    1    VDJ_cgene
#> 26                    1    VDJ_cgene
#> 27                    1    VDJ_cgene
#> 28                    1    VDJ_cgene
#> 29                    1    VDJ_cgene
#> 30                    1    VDJ_cgene
#> 31                    1    VDJ_cgene
#> 32                    1    VDJ_cgene
#> 33                    1    VDJ_cgene