Makes a Circos plot from the VDJ_analyze output. Connects the V-alpha with the corresponding V-beta gene for each clonotype.

Makes a Circos plot from the VDJ_analyze output. Connects the V-alpha with the corresponding V-beta gene for each clonotype.

VDJ_alpha_beta_Vgene_circos(
  VDJ,
  V.or.J,
  B.or.Tcells,
  label.threshold,
  c.threshold,
  cell.level,
  clonotype.per.gene.threshold,
  c.count,
  platypus.version,
  filter1H1L
)

Arguments

VDJ
The output of the VDJ_GEX_integrate function (Platypus platypus.version v2). A list of data frames for each sample containing the clonotype information and cluster membership information. For Platypus platypus.version v3, VDJ_GEX_matrix.output[[1]] has to be supplied.

V.or.J
Determines whether to plot the alpha beta gene pairing of the V or J genes. "V", "J" or "both" as possible inputs. Default: "both".

B.or.Tcells
Specify whether B or T cells are being analyzed ("B" or "T"). If not specified, function attempts to decide based on gene names.

label.threshold
Minimal amount of clonotypes per gene neccessary to add a gene label to the sector. Default: 0.

c.threshold
Only clonotypes are considered with a frequency higher then c.threshold. Allows to filter for only highly expanded clonotypes.

cell.level
Logical, defines whether weight of connection should be based on number of clonotypes of number of cells. Default: number of clonotypes.

clonotype.per.gene.threshold
How many clonotypes are required to plot a sector for a gene. Filters the rows and colums of the final adjacency matrix.

c.count
Show clonotype or cell count on Circos plot. Default = T.

platypus.version
Which platypus.version of platypus is beeing used. Default = v2.

filter1H1L
Whether to filter the input VDJ.matrix in "v3" to only include cells with 1 VDJ and 1 VJ chain. Defaults to TRUE

Value

Returns list of plots. The first n elements contain the circos plot of the n datasets from the VDJ.analyze function. The n+1 element contains a list of the n adjancey matrices for each dataset.

Examples

if (FALSE) {
 plots <- VDJ_alpha_beta_Vgene_circos(VDJ_GEX_matrix.output[[1]])
}