Calculates and plots common diversity and overlap measures for repertoires and alike. Require the vegan package

Calculates and plots common diversity and overlap measures for repertoires and alike. Require the vegan package

VDJ_diversity(
  VDJ,
  feature.columns,
  grouping.column,
  metric,
  subsample.to.same.n,
  pvalues.label.size,
  axis.label.size,
  platypus.version
)

Arguments

VDJ
VDJ dataframe output from either the VDJ_analyse (platypus.version = "v2") or from the VDJ_GEX_matrix function (platypus.version = "v3")(VDJ_GEX_matrix.output[[1]])

feature.columns
Character vector. One or more column names from the VDJ of which diversity or overlap metrics are calculated. if more than one column is provided (e.g. c("VDJ_cdr3s_aa","VJ_cdr3s_aa")) these columns will be pasted together before metric calculation. Defaults to "CDRH3_aa" if platypus.version == "v2" and "VDJ_cdr3s_aa" if platypus.version == "v3".

grouping.column
Character. Column name of a column to group metrics by. This could be "sample_id" to calculate the metric for each sample. This column is required if metric = "simpson". If so, the simpson overlap index will be calculated pairwise for all combinations of elements in the grouping.column. Defaults to "none".

metric
Character. Diversity or overlap metric to calculate. Can be c("richness", "bergerparker", "simpson", "ginisimpson", "shannon", "shannonevenness", "jaccard"). Defaults to "shannon". If jaccard is selected, a heatmap with the pairwise comparisons between all groups is returned. If any of the others is selected, a dotplot is returned

subsample.to.same.n
Boolean defaults to TRUE. Whether to subsample larger groups down to the size of the smallest group

pvalues.label.size
Numeric. Only used if overlap indices are calculated. Defaults to 4. Is passed on to ggplot theme

axis.label.size
Numeric. Only used if overlap indices are calculated. Defaults to 12. Is passed on to ggplot theme

platypus.version
Version of platypus to use. Defaults to "v2". If an output of the VDJ_analyze function is supplied, set to "v2". If an output of the VDJ_GEX_matrix function is supplied set to "v3"

Value

Returns a ggplot with the calculated metric for each group (if provided). Data is accessible via ggplot.output$data

Examples