Expand the aberrant cells in a VDJ dataframe by converting them into additional rows

Expand the aberrant cells in a VDJ dataframe by converting them into additional rows. Aberrant cells consist of cells with more than 1 VDJ or VJ chain.

VDJ_expand_aberrants(
  VDJ,
  chain.to.expand,
  add.barcode.prefix,
  additional.VDJ.features,
  additional.VJ.features,
  add.CDR3aa,
  add.expanded.number,
  recalculate.clonotype.frequency
)

Arguments

VDJ

VDJ or VDJ.GEX.matrix[[1]] object, as obtained from the VDJ_GEX_matrix function in Platypus.

chain.to.expand

string, 'VDJ' to expand VDJ aberrants, 'VJ' to expand VJ aberrants, 'VDJ.VJ' for both.

add.barcode.prefix

boolean - if T, a new barcode will be added for each expanded aberrant.

additional.VDJ.features

vector of strings - VDJ_expand_aberrants will only expand across the sequence columns of VDJ. If you have additional columns with aberrant cell features (e.g., both 'yes' and 'no' binders for a single sequence), where the aberrants are VDJ-specific, include them here.

additional.VJ.features

vector of strings - VDJ_expand_aberrants will only expand across the sequence columns of VDJ. If you have additional columns with aberrant cell features (e.g., both 'yes' and 'no' binders for a single sequence), where the aberrants are VJ-specific, include them here.

add.CDR3aa

boolean - if T, will create a new column 'CDR3aa' with pasted VDJ_cdr3s_aa and VJ_cdr3s_aa.

add.expanded.number

boolean - if T, will add the number of new cells resulting from an aberrant one.

recalculate.clonotype.frequency

boolean - if T, will recalculate the clonotype frequencies for the resulting, expanded VDJ.

Value

Returns a VDJ format dataframe in which cells with more than one VDJ or VJ chain are split into multiple rows each containing only one VDJ VJ chain combination.

Examples

VDJ_expand_aberrants(VDJ = small_vgm[[1]],
chain.to.expand='VDJ.VJ',
add.barcode.prefix=TRUE, recalculate.clonotype.frequency=FALSE)
#>                    barcode sample_id group_id clonotype_id_10x celltype
#> 1    s1_TATGCCCCAGCTTCGG-1        s1     rMOG       clonotype7   B cell
#> 2    s1_CCCAGTTAGCAGGTCA-1        s1     rMOG    clonotype2053   B cell
#> 3    s1_CGTTGGGAGGGATCTG-1        s1     rMOG    clonotype2386   B cell
#> 4    s2_CGAACATAGTGACATA-1        s2     rMOG     clonotype441   B cell
#> 5    s1_ACGATGTTCAAACCGT-1        s1     rMOG    clonotype2040   B cell
#> 6  s2_1_AGGTCATTCTTGGGTA-1        s2     rMOG     clonotype516   B cell
#> 7  s2_2_AGGTCATTCTTGGGTA-1        s2     rMOG     clonotype516   B cell
#> 8    s3_ATGCGATTCCCATTAT-1        s3 MOG35-55     clonotype141   B cell
#> 9    s1_ATGAGGGCAATGAATG-1        s1     rMOG     clonotype743   B cell
#> 10   s3_AGCGTATCACGGACAA-1        s3 MOG35-55     clonotype390   B cell
#> 11   s3_GTGCGGTAGCCCAGCT-1        s3 MOG35-55     clonotype464   B cell
#> 12   s3_TTCTTAGCATTAGGCT-1        s3 MOG35-55      clonotype10   B cell
#> 13   s1_GCTGGGTGTCCAGTTA-1        s1     rMOG    clonotype1725   B cell
#> 14   s1_CTACGTCCAAAGGCGT-1        s1     rMOG    clonotype1243   B cell
#> 15   s1_GATCGATAGTTAGCGG-1        s1     rMOG    clonotype2271   B cell
#> 16   s3_TGGGCGTTCGGTTCGG-1        s3 MOG35-55    clonotype1028   B cell
#> 17   s3_ATTATCCAGTTCGATC-1        s3 MOG35-55     clonotype592   B cell
#> 18   s2_CCTCTGACACCAGATT-1        s2     rMOG     clonotype649   B cell
#> 19   s2_CTACCCAAGACTAGAT-1        s2     rMOG     clonotype634   B cell
#>    Nr_of_VDJ_chains Nr_of_VJ_chains         VDJ_cdr3s_aa
#> 1                 0               1                     
#> 2                 1               0     CARSSLYYGNFYFDYW
#> 3                 1               1 CTRNMRLRRGTGTGYAMDYW
#> 4                 0               1                     
#> 5                 1               1     CVRGGYDYDRDYFDFW
#> 6                 2               2     CARNRITTVVAPMDYW
#> 7                 2               2         CARKDYARGDYW
#> 8                 1               1   CARWGIYDGYYGDAMDYW
#> 9                 0               1                     
#> 10                1               1             CARTFAYW
#> 11                1               1           CARMVTGAYW
#> 12                0               1                     
#> 13                1               1         CARVNLHAMDYW
#> 14                1               1             CVRSWDYW
#> 15                1               1   CTFRIYYYGSSPYYFDYW
#> 16                1               1    CAREGYYGSSPYAMDYW
#> 17                1               1         CAREATVVADYW
#> 18                1               1       CASGGGNPWYFDVW
#> 19                1               1         CARPPLTGFAYW
#>                VJ_cdr3s_aa
#> 1              CFQGSHVPWTF
#> 2                         
#> 3              CLQYDNLLWTF
#> 4              CAQNLELPWTF
#> 5              CAQNLELPYTF
#> 6  CQQSNSWPLTF;CQQGSSIPLTF
#> 7  CQQSNSWPLTF;CQQGSSIPLTF
#> 8              CALWYSNHLVF
#> 9              CLQYDEFPSTF
#> 10             CLQYASSPWTF
#> 11             CQNDYSYPLTF
#> 12         CGVGDTIKEQFVYVF
#> 13             CQHFWGTPRTF
#> 14             CQQYNSYPLTF
#> 15             CQQHYSTPFTF
#> 16              CLQYDNLFTF
#> 17             CQQSNEDPRTF
#> 18             CQQHYSTPYTF
#> 19              CHQYLSSWTF
#>                                                    VDJ_cdr3s_nt
#> 1                                                              
#> 2              TGTGCAAGATCTTCACTCTACTATGGTAACTTCTACTTTGACTACTGG
#> 3  TGTACAAGAAATATGAGATTACGACGCGGGACTGGGACTGGGTATGCTATGGACTACTGG
#> 4                                                              
#> 5              TGTGTCAGGGGGGGGTATGATTACGACAGGGACTACTTTGACTTCTGG
#> 6              TGTGCCAGAAATCGTATTACTACGGTAGTAGCCCCTATGGACTACTGG
#> 7                          TGTGCCAGAAAAGACTACGCGAGAGGGGACTACTGG
#> 8        TGTGCAAGATGGGGGATCTATGATGGTTACTACGGGGATGCTATGGACTACTGG
#> 9                                                              
#> 10                                     TGTGCAAGGACGTTTGCTTACTGG
#> 11                               TGTGCCCGTATGGTTACGGGTGCTTACTGG
#> 12                                                             
#> 13                         TGTGCAAGAGTTAACCTGCATGCTATGGACTACTGG
#> 14                                     TGTGTGAGAAGTTGGGACTACTGG
#> 15       TGTACTTTCCGAATTTATTACTACGGTAGTAGCCCTTACTACTTTGACTACTGG
#> 16          TGTGCAAGAGAGGGTTACTACGGTAGTAGTCCCTATGCTATGGACTACTGG
#> 17                         TGTGCAAGAGAGGCTACGGTAGTAGCGGACTACTGG
#> 18                   TGTGCAAGCGGAGGGGGTAACCCCTGGTACTTCGATGTCTGG
#> 19                         TGTGCTCGACCCCCTCTAACTGGTTTTGCTTACTGG
#>                                                            VJ_cdr3s_nt
#> 1                                    TGCTTTCAAGGTTCACATGTTCCGTGGACGTTC
#> 2                                                                     
#> 3                                    TGTCTACAGTATGATAATCTTCTGTGGACGTTC
#> 4                                    TGTGCTCAAAATCTAGAACTTCCGTGGACGTTC
#> 5                                    TGTGCTCAAAATCTTGAACTTCCGTACACGTTC
#> 6  TGTCAACAAAGTAATAGCTGGCCGCTCACGTTC;TGCCAGCAGGGTAGTAGTATACCGCTCACGTTC
#> 7  TGTCAACAAAGTAATAGCTGGCCGCTCACGTTC;TGCCAGCAGGGTAGTAGTATACCGCTCACGTTC
#> 8                                    TGTGCTCTATGGTACAGCAACCATTTGGTGTTC
#> 9                                    TGTCTACAGTATGATGAGTTTCCGTCCACGTTC
#> 10                                   TGTCTACAATATGCTAGTTCTCCGTGGACGTTC
#> 11                                   TGTCAGAATGATTATAGTTATCCGCTCACGTTC
#> 12                       TGTGGTGTGGGTGATACAATTAAGGAACAATTTGTGTATGTTTTC
#> 13                                   TGTCAACATTTTTGGGGTACTCCTCGGACGTTC
#> 14                                   TGTCAGCAATATAACAGCTATCCGCTCACGTTC
#> 15                                   TGTCAGCAACATTATAGTACTCCATTCACGTTC
#> 16                                      TGTCTACAGTATGATAATCTATTCACGTTC
#> 17                                   TGTCAGCAAAGTAATGAGGATCCTCGGACGTTC
#> 18                                   TGTCAGCAACATTATAGTACTCCGTACACGTTC
#> 19                                      TGTCATCAATACCTCTCCTCGTGGACGTTC
#>               VDJ_chain_contig
#> 1                             
#> 2  CCCAGTTAGCAGGTCA-1_contig_1
#> 3  CGTTGGGAGGGATCTG-1_contig_2
#> 4                             
#> 5  ACGATGTTCAAACCGT-1_contig_1
#> 6  AGGTCATTCTTGGGTA-1_contig_3
#> 7  AGGTCATTCTTGGGTA-1_contig_4
#> 8  ATGCGATTCCCATTAT-1_contig_2
#> 9                             
#> 10 AGCGTATCACGGACAA-1_contig_1
#> 11 GTGCGGTAGCCCAGCT-1_contig_2
#> 12                            
#> 13 GCTGGGTGTCCAGTTA-1_contig_2
#> 14 CTACGTCCAAAGGCGT-1_contig_1
#> 15 GATCGATAGTTAGCGG-1_contig_1
#> 16 TGGGCGTTCGGTTCGG-1_contig_2
#> 17 ATTATCCAGTTCGATC-1_contig_2
#> 18 CCTCTGACACCAGATT-1_contig_2
#> 19 CTACCCAAGACTAGAT-1_contig_2
#>                                            VJ_chain_contig VDJ_chain VJ_chain
#> 1                              TATGCCCCAGCTTCGG-1_contig_1                IGK
#> 2                                                                IGH         
#> 3                              CGTTGGGAGGGATCTG-1_contig_1       IGH      IGK
#> 4                              CGAACATAGTGACATA-1_contig_1                IGK
#> 5                              ACGATGTTCAAACCGT-1_contig_2       IGH      IGK
#> 6  AGGTCATTCTTGGGTA-1_contig_1;AGGTCATTCTTGGGTA-1_contig_2       IGH  IGK;IGK
#> 7  AGGTCATTCTTGGGTA-1_contig_1;AGGTCATTCTTGGGTA-1_contig_2       IGH  IGK;IGK
#> 8                              ATGCGATTCCCATTAT-1_contig_1       IGH      IGL
#> 9                              ATGAGGGCAATGAATG-1_contig_1                IGK
#> 10                             AGCGTATCACGGACAA-1_contig_2       IGH      IGK
#> 11                             GTGCGGTAGCCCAGCT-1_contig_1       IGH      IGK
#> 12                             TTCTTAGCATTAGGCT-1_contig_1                IGL
#> 13                             GCTGGGTGTCCAGTTA-1_contig_1       IGH      IGK
#> 14                             CTACGTCCAAAGGCGT-1_contig_2       IGH      IGK
#> 15                             GATCGATAGTTAGCGG-1_contig_2       IGH      IGK
#> 16                             TGGGCGTTCGGTTCGG-1_contig_1       IGH      IGK
#> 17                             ATTATCCAGTTCGATC-1_contig_1       IGH      IGK
#> 18                             CCTCTGACACCAGATT-1_contig_1       IGH      IGK
#> 19                             CTACCCAAGACTAGAT-1_contig_1       IGH      IGK
#>    VDJ_vgene          VJ_vgene VDJ_dgene VDJ_jgene    VJ_jgene VDJ_cgene
#> 1                    IGKV1-117                           IGKJ1          
#> 2   IGHV1-81                     IGHD2-8     IGHJ2                IGHG2C
#> 3  IGHV5-9-1         IGKV19-93               IGHJ4       IGKJ1    IGHG2C
#> 4                    IGKV2-109                           IGKJ1          
#> 5    IGHV2-2         IGKV2-109               IGHJ2       IGKJ2      IGHM
#> 6  IGHV2-9-1 IGKV5-48;IGKV4-91               IGHJ4 IGKJ5;IGKJ5      IGHM
#> 7    IGHV2-2 IGKV5-48;IGKV4-91               IGHJ4 IGKJ5;IGKJ5      IGHM
#> 8   IGHV1-69             IGLV1   IGHD2-3     IGHJ4       IGLJ1      IGHM
#> 9                   IGKV14-111                           IGKJ2          
#> 10  IGHV1-64         IGKV9-120               IGHJ3       IGKJ1      IGHM
#> 11  IGHV1-64          IGKV8-19               IGHJ3       IGKJ5      IGHM
#> 12                       IGLV3                           IGLJ2          
#> 13   IGHV7-1         IGKV12-46               IGHJ4       IGKJ1    IGHG2C
#> 14  IGHV10-1          IGKV6-15               IGHJ2       IGKJ5      IGHM
#> 15  IGHV14-4          IGKV6-17   IGHD1-1     IGHJ2       IGKJ4      IGHD
#> 16  IGHV1-55         IGKV19-93               IGHJ4       IGKJ4      IGHD
#> 17   IGHV9-3           IGKV3-5               IGHJ2       IGKJ1      IGHM
#> 18  IGHV1-81          IGKV6-17               IGHJ1       IGKJ2      IGHM
#> 19   IGHV8-8          IGKV8-27   IGHD4-1     IGHJ3       IGKJ1      IGHM
#>     VJ_cgene
#> 1       IGKC
#> 2           
#> 3       IGKC
#> 4       IGKC
#> 5       IGKC
#> 6  IGKC;IGKC
#> 7  IGKC;IGKC
#> 8      IGLC1
#> 9       IGKC
#> 10      IGKC
#> 11      IGKC
#> 12     IGLC2
#> 13      IGKC
#> 14      IGKC
#> 15      IGKC
#> 16      IGKC
#> 17      IGKC
#> 18      IGKC
#> 19      IGKC
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                     VDJ_sequence_nt_raw
#> 1                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                      
#> 2                                                                                                                                 ACAACCTATGATCAATGTCTTCTTCACAGTCCCTGAACACACTGACTCTAACCATGGAATGGATCTGGATCTTTCTCTTCATCCTGTCAGGAACTGCAGGTGTCCAATCCCAGGTTCAGCTGCAGCAGTCTGGAGCTGAGCTGGCGAGGCCTGGGGCTTCAGTGAAGCTGTCCTGCAAGGCTTCTGGCTACACCTTCACAAGCTATGGTATAAGCTGGGTGAAGCAGAGAACTGGACAGGGCCTTGAGTGGATTGGAGAGATTTATCCTAGAAGTGGTAATACTTACTACAATGAGAAGTTCAAGGGCAAGGCCACACTGACTGCAGACAAATCCTCCAGCACAGCGTACATGGAGCTCCGCAGCCTGACATCTGAGGACTCTGCGGTCTATTTCTGTGCAAGATCTTCACTCTACTATGGTAACTTCTACTTTGACTACTGGGGCCAAGGCACCACTCTCACAGTCTCCTCAGCCAAAACAACAGCCCCATCGGTCTATCCACTGGCCCCTGTGTGTGGAGGTACAACTGGCTCCTCGGTGACTCTAGGATGCCTGGTCAAGGG
#> 3                                                                                                              CTGGAATTGATTCCTAGTTCCTCACGTTCAGTGATGAGTACTGAACACAGACCCCTCACCATGAACTTCGGGCTCAGATTGATTTTCCTTGTCCTTACTTTAAAAGGTGTCCAGTGTGACGTGAAGCTGGTGGAGTCTGGGGAAGGCTTAGTGAAGCCTGGAGGGTCCCTGAAACTCTCCTGTGCAGCCTCTGGATTCACTTTCAGTAGCTATGCCATGTCTTGGGTTCGCCAGACTCCAGAGAAGAGGCTGGAGTGGGTCGCATACATTAGTAGTGGTGGTGATTACATCTACTATGCAGACACTGTGAAGGGCCGATTCACCATCTCCAGAGACAATGCCAGGAACACCCTGTACCTGCAAATGAGCAGTCTGAAGTCTGAGGACACAGCCATGTATTACTGTACAAGAAATATGAGATTACGACGCGGGACTGGGACTGGGTATGCTATGGACTACTGGGGTCAAGGAACCTCAGTCACCGTCTCCTCAGCCAAAACAACAGCCCCATCGGTCTATCCACTGGCCCCTGTGTGTGGAGGTACAACTGGCTCCTCGGTGACTCTAGGATGCCTGGTCAAGGG
#> 4                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                      
#> 5                                                                                                                                               GATCCTCTTCTCATAGAGCCTCCATCAGAGCATGGCTGTCTTGGGGCTGCTCTTCTGCCTGGTGACATTCCCAAGCTGTGTCCTATCCCAGGTGCAGCTGAAGCAGTCAGGACCTGGCCTAGTGCAGCCCTCACAGAGCCTGTCCATCACCTGCACAGTCTCTGGTTTCTCATTAACTAGCTATGGTGTACACTGGATTCGCCAGTCTCCAGGAAAGGGTCTGGAGTGGCTGGGAGTGATATGGAGTGGTGGAACCACAGACTATAATGCAGCTTTCATATCCAGACTGAGCATCAGTAAGGACAGATCCAAGAGCCAAGTTTTCTTTAAAATGAACAGTCTGCAAGTTGATGACACAGCCATATATTATTGTGTCAGGGGGGGGTATGATTACGACAGGGACTACTTTGACTTCTGGGGCCAAGGCACCACTCTCTCAGTCTCCTCAGAGAGTCAGTCCTTCCCAAATGTCTTCCCCCTCGTCTCCTGCGAGAGCCCCCTGTCTGATAAGAATCTGGTGGCCATGGGCTGCCTGGCCCGGGACTTCCTGC
#> 6                              TGGGGATCCTCTTCTCATAGAGCCTCCATCAGAGCATGGCTGTCCTGGCGCTACTCCTCTGCCTGGTGACTTTCCCAAGCTGTGCCCTGTCCCAGGTGCAGCTGAAGGAGTCAGGACCTGGCCTGGTGGCGCCCTCACAGAGCCTGTCCATCACATGCACTGTCTCTGGGTTCTCATTAACCAGCTATGCTATAAGCTGGGTTCGCCAGCCACCAGGAAAGGGTCTGGAGTGGCTTGGAGTAATATGGACTGGTGGAGGCACAAATTATAATTCAGCTCTCAAATCCAGACTGAGCATCAGCAAAGACAACTCCAAGAGTCAAGTTTTCTTAAAAATGAACAGTCTGCAAACTGATGACACAGCCAGGTACTACTGTGCCAGAAATCGTATTACTACGGTAGTAGCCCCTATGGACTACTGGGGTCAAGGAACCTCAGTCACCGTCTCCTCAGAGAGTCAGTCCTTCCCAAATGTCTTCCCCCTCGTCTCCTGCGAGAGCCCCCTGTCTGATAAGAATCTGGTGGCCATGGGCTGCCTGGCCCGGGACTTCCTGCCCAGCACCATTTCCTTCACCTGGAACTACCAGAACAACACTGAAGTCATCCAGGGTATCAGAACCTTCCCAACACTGAGGACAGGGGGCAAGTACCTAGCCACCTCGCA
#> 7                                          TGGGGATCCTCTTCTCATAGAGCCTCCATCAGAGCATGGCTGTCTTGGGGCTGCTCTTCTGCCTGGTGACATTCCCAAGCTGTGTCCTATCCCAGGTGCAGCTGAAGCAGTCAGGACCTGGCCTAGTGCAGCCCTCACAGAGCCTGTCCATCACCTGCACAGTCTCTGGTTTCTCATTAACTAGCTATGGTGTACACTGGGTTCGCCAGTCTCCAGGAAAGGGTCTGGAGTGGCTGGGAGTGATATGGAGTGGTGGAAGCACAGACTATAATGCAGCTTTCATATCCAGACTGAGCATCAGCAAGGACAATTCCAAGAGCCAAGTTTTCTTTAAAATGAACAGTCTGCAAGCTGATGACACAGCCATATATTACTGTGCCAGAAAAGACTACGCGAGAGGGGACTACTGGGGTCAAGGAACCTCAGTCACCGTCTCCTCAGAGAGTCAGTCCTTCCCAAATGTCTTCCCCCTCGTCTCCTGCGAGAGCCCCCTGTCTGATAAGAATCTGGTGGCCATGGGCTGCCTGGCCCGGGACTTCCTGCCCAGCACCATTTCCTTCACCTGGAACTACCAGAACAACACTGAAGTCATCCAGGGTATCAGAACCTTCCCAACACTGAGGACAGGGGGCAAGTACCTAGCCACCTCGCA
#> 8    GAGCATAAGATCACTGTTCTCTCTACAGTTACTGAGCACACAGGACCTCACCATGGGATGGAGCTGTATCATCCTCTTCTTGGTATCAACAGCTACAGGTGTCCACTCCCAGGTCCAACTGCAGCAGCCTGGGGCTGAGCTTGTGATGCCTGGGGCTTCAGTGAAGCTGTCCTGCAAGGCTTCTGGCTACACCTTCACCAGCTACTGGATGCACTGGGTGAAGCAGAGGCCTGGACAAGGCCTTGAGTGGATCGGAGAGATTGATCCTTCTGATAGTTATACTAACTACAATCAAAAGTTCAAGGGCAAGTCCACATTGACTGTAGACAAATCCTCCAGCACAGCCTACATGCAGCTCAGCAGCCTGACATCTGAGGACTCTGCGGTCTATTACTGTGCAAGATGGGGGATCTATGATGGTTACTACGGGGATGCTATGGACTACTGGGGTCAAGGAACCTCAGTCACCGTCTCCTCAGAGAGTCAGTCCTTCCCAAATGTCTTCCCCCTCGTCTCCTGCGAGAGCCCCCTGTCTGATAAGAATCTGGTGGCCATGGGCTGCCTGGCCCGGGACTTCCTGCCCAGCACCATTTCCTTCACCTGGAACTACCAGAACAACACTGAAGTCATCCAGGGTATCAGAACCTTCCCAACACTGAGGACAGGGGGCAAGTACCTAGCCACCTCGCA
#> 9                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                      
#> 10                                               AGTTCTCTCTACAGTTACTGAGCACACAGGACCTCACAATGGGATGGAGCTATATCATCCTCTTTTTGGTAGCAACAGCTACAGGTGTCCACTCCCAGGTCCAACTGCAGCAGCCTGGGGCTGAGCTGGTAAAGCCTGGGGCTTCAGTGAAGTTGTCCTGCAAGGCTTCTGGCTACACTTTCACCAGCTACTGGATGCACTGGGTGAAGCAGAGGCCTGGACAAGGCCTTGAGTGGATTGGAATGATTCATCCTAATAGTGGTAGTACTAACTACAATGAGAAGTTCAAGAGCAAGGCCACACTGACTGTAGACAAATCCTCCAGCACAGCCTACATGCAACTCAGCAGCCTGACATCTGAGGACTCTGCGGTCTATTACTGTGCAAGGACGTTTGCTTACTGGGGCCAAGGGACTCTGGTCACTGTCTCTGCAGAGAGTCAGTCCTTCCCAAATGTCTTCCCCCTCGTCTCCTGCGAGAGCCCCCTGTCTGATAAGAATCTGGTGGCCATGGGCTGCCTGGCCCGGGACTTCCTGCCCAGCACCATTTCCTTCACCTGGAACTACCAGAACAACACTGAAGTCATCCAGGGTATCAGAACCTTCCCAACACTGAGGACAGGGGGCAAGTACCTAGCCACCTCGCA
#> 11                         AAAAACATGAGATCACAGTTCTCTCTACAGTTACTGAGCACACAGGACCTCACAATGGGATGGAGCTATATCATCCTCTTTTTGGTAGCAACAGCTACAGGTGTCCACTCCCAGGTCCAACTGCAGCAGCCTGGGGCTGAGCTGGTAAAGCCTGGGGCTTCAGTGAAGTTGTCCTGCAAGGCTTCTGGCTACACTTTCACCAGCTACTGGATGCACTGGGTGAAGCAGAGGCCTGGACAAGGCCTTGAGTGGATTGGAATGATTCATCCTAATAGTGGTAGTACTAACTACAATGAGAAGTTCAAGAGCAAGGCCACACTGACTGTAGACAAATCCTCCAGCACAGCCTACATGCAACTCAGCAGCCTGACATCTGAGGACTCTGCGGTCTATTACTGTGCCCGTATGGTTACGGGTGCTTACTGGGGCCAAGGGACTCTGGTCACTGTCTCTGCAGAGAGTCAGTCCTTCCCAAATGTCTTCCCCCTCGTCTCCTGCGAGAGCCCCCTGTCTGATAAGAATCTGGTGGCCATGGGCTGCCTGGCCCGGGACTTCCTGCCCAGCACCATTTCCTTCACCTGGAACTACCAGAACAACACTGAAGTCATCCAGGGTATCAGAACCTTCCCAACACTGAGGACAGGGGGCAAGTACCTAGCCACCTCGCA
#> 12                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                     
#> 13                                                                                                                      TGGGGAGTGGGATCCCGTCCTGAGTTCCCCAATCTTCACATTCAGAAATCACCACTCAGTCCTGTCACTATGAAGTTGTGGTTAAACTGGGTTTTTCTTTTAACACTTTTACATGGTATCCAGTGTGAGGTGAAGCTGGTGGAATCTGGAGGAGGCTTGGTACAGTCTGGGCGTTCTCTGAGACTCTCCTGTGCAACTTCTGGGTTCACCTTCAGTGATTTCTACATGGAGTGGGTCCGCCAAGCTCCAGGGAAGGGACTGGAGTGGATTGCTGCAAGTAGAAACAAAGCTAATGATTATACAACAGAGTACAGTGCATCTGTGAAGGGTCGGTTCATCGTCTCCAGAGACACTTCCCAAAGCATCCTCTACCTTCAGATGAATGCCCTGAGAGCTGAGGACACTGCCATTTATTACTGTGCAAGAGTTAACCTGCATGCTATGGACTACTGGGGTCAAGGAACCTCAGTCACCGTCTCCTCAGCCAAAACAACAGCCCCATCGGTCTATCCACTGGCCCCTGTGTGTGGAGGTACAACTGGCTCCTCGGTGACTCTAGGATGCCTGGTCAAGGG
#> 14               GAGGCAGAGAACTTTAGCCCTGTCTTCTTTTTTAGTGTTCAGCACTGACAATATGACATTGAACATGCTGTTGGGGCTGAAGTGGGTTTTCTTTGTTGTTTTTTATCAAGGTGTGCATTGTGAGGTGCAGCTTGTTGAGTCTGGTGGAGGATTGGTGCAGCCTAAAGGGTCATTGAAACTCTCATGTGCAGCCTCTGGATTCAGCTTCAATACCTACGCCATGAACTGGGTCCGCCAGGCTCCAGGAAAGGGTTTGGAATGGGTTGCTCGCATAAGAAGTAAAAGTAATAATTATGCAACATATTATGCCGATTCAGTGAAAGACAGATTCACCATCTCCAGAGATGATTCAGAAAGCATGCTCTATCTGCAAATGAACAACTTGAAAACTGAGGACACAGCCATGTATTACTGTGTGAGAAGTTGGGACTACTGGGGCCAAGGCACCACTCTCACAGTCTCCTCAGAGAGTCAGTCCTTCCCAAATGTCTTCCCCCTCGTCTCCTGCGAGAGCCCCCTGTCTGATAAGAATCTGGTGGCCATGGGCTGCCTGGCCCGGGACTTCCTGCCCAGCACCATTTCCTTCACCTGGAACTACCAGAACAACACTGAAGTCATCCAGGGTATCAGAACCTTCCCAACACTGAGGACAGGGGGCAAGTACCTAGCCACCTCGCA
#> 15 GGGGACATATGAACACTGTTTTCTCTACAGTCACTGAATCTCAATGTCCTTACAATGAAATGCAGCTGGGTCATCTTCTTCCTGATGGCAGTGGTTATAGGGGTCAATTCAGAGGTTCAGCTGCAGCAGTCTGGGGCTGAGCTTGTGAGGCCAGGGGCCTCAGTCAAGTTGTCCTGCACAGCTTCTGGCTTTAACATTAAAGACGACTATATGCACTGGGTGAAGCAGAGGCCTGAACAGGGCCTGGAGTGGATTGGATGGATTGATCCTGAGAATGGTGATACTGAATATGCCTCGAAGTTCCAGGGCAAGGCCACTATAACAGCAGACACATCCTCCAACACAGCCTACCTGCAGCTCAGCAGCCTGACATCTGAGGACACTGCCGTCTATTACTGTACTTTCCGAATTTATTACTACGGTAGTAGCCCTTACTACTTTGACTACTGGGGCCAAGGCACCACTCTCACAGTCTCCTCAGGTAATGAAAAGGGACCTGACATGTTCCTCCTCTCAGAGTGCAAAGCCCCAGAGGAAAATGAAAAGATAAACCTGGGCTGTTTAGTAATTGGAAGTCAGCCACTGAAAATCAGCTGGGAGCCAAAGAAGTCAAGTATAGTTGAACATGTCTTCCCCTCTGAAATGAGAAATGGCAATTATACAATGGTCCTCCAGGTCACTGTGCTGGCCTC
#> 16      AAGCATAAGATCACTGTTCTCTCTACAGTTACTAAGCACACAGGATCTCACCATGGGATGGAGCTGTATCATCCTCATTTTGGTAGCAGCAGCTACAGGTGTCCACTCCCAGGTCCAACTGCAGCAGCCTGGGGCTGAGCTTGTGAAGCCTGGGGCTTCAGTGAAGATGTCCTGCAAGGCTTCTGGCTACACCTTCACCAGCTACTGGATAACCTGGGTGAAGCAGAGGCCTGGACAAGGCCTTGAGTGGATTGGAGATATTTATCCTGGTAGTGGTAGTACTAACTACAATGAGAAGTTCAAGAGCAAGGCCACACTGACTGTAGACACATCCTCCAGCACAGCCTACATGCAGCTCAGCAGCCTGACATCTGAGGACTCTGCGGTCTATTACTGTGCAAGAGAGGGTTACTACGGTAGTAGTCCCTATGCTATGGACTACTGGGGTCAAGGAACCTCAGTCACCGTCTCCTCAGGTAATGAAAAGGGACCTGACATGTTCCTCCTCTCAGAGTGCAAAGCCCCAGAGGAAAATGAAAAGATAAACCTGGGCTGTTTAGTAATTGGAAGTCAGCCACTGAAAATCAGCTGGGAGCCAAAGAAGTCAAGTATAGTTGAACATGTCTTCCCCTCTGAAATGAGAAATGGCAATTATACAATGGTCCTCCAGGTCACTGTGCTGGCCTC
#> 17               TGGGGAAGGGAGTGACCAGTTAGTCTTAAGGCACCACTGAGCCCAAGTCTTAGACATCATGGGTTGGCTGTGGAACTTGCTATTCCTGATGGCAGCTGCCCAAAGTGCCCAAGCACAGATCCAGTTGGTACAGTCTGGACCTGAGCTGAAGAAGCCTGGAGAGACAGTCAAGATCTCCTGCAAGGCTTCTGGGTATACCTTCACAACCTATGGAATGAGCTGGGTGAAACAGGCTCCAGGAAAGGGTTTAAAGTGGATGGGCTGGATAAACACCTACTCTGGAGTGCCAACATATGCTGATGACTTCAAGGGACGGTTTGCCTTCTCTTTGGAAACCTCTGCCAGCACTGCCTATTTGCAGATCAACAACCTCAAAAATGAGGACACGGCTACATATTTCTGTGCAAGAGAGGCTACGGTAGTAGCGGACTACTGGGGCCAAGGCACCACTCTCACAGTCTCCTCAGAGAGTCAGTCCTTCCCAAATGTCTTCCCCCTCGTCTCCTGCGAGAGCCCCCTGTCTGATAAGAATCTGGTGGCCATGGGCTGCCTGGCCCGGGACTTCCTGCCCAGCACCATTTCCTTCACCTGGAACTACCAGAACAACACTGAAGTCATCCAGGGTATCAGAACCTTCCCAACACTGAGGACAGGGGGCAAGTACCTAGCCACCTCGCA
#> 18                                                 GGGACACTGACTCTAACCATGGAATGGATCTGGATCTTTCTCTTCATCCTGTCAGGAACTGCAGGTGTCCAATCCCAGGTTCAGCTGCAGCAGTCTGGAGCTGAGCTGGCGAGGCCTGGGGCTTCAGTGAAGCTGTCCTGCAAGGCTTCTGGCTACACCTTCACAAGCTATGGTATAAGCTGGGTGAAGCAGAGAACTGGACAGGGCCTTGAGTGGATTGGAGAGATTTATCCTAGAAGTGGTAATACTTACTACAATGAGAAGTTCAAGGGCAAGGCCACACTGACTGCAGACAAATCCTCCAGCACAGCGTACATGGAGCTCCGCAGCCTGACATCTGAGGACTCTGCGGTCTATTTCTGTGCAAGCGGAGGGGGTAACCCCTGGTACTTCGATGTCTGGGGCACAGGGACCACGGTCACCGTCTCCTCAGAGAGTCAGTCCTTCCCAAATGTCTTCCCCCTCGTCTCCTGCGAGAGCCCCCTGTCTGATAAGAATCTGGTGGCCATGGGCTGCCTGGCCCGGGACTTCCTGCCCAGCACCATTTCCTTCACCTGGAACTACCAGAACAACACTGAAGTCATCCAGGGTATCAGAACCTTCCCAACACTGAGGACAGGGGGCAAGTACCTAGCCACCTCGCA
#> 19                                                      GGGGAAGTGTGCAGCCATGGGCAGGCTTACTTCTTCATTCCTGTTACTGATTGTCCCTGCATATGTCCTGTCCCAGGTTACTCTGAAAGAGTCTGGCCCTGGGATATTGCAGCCCTCCCAGACCCTCAGTCTGACTTGTTCTTTCTCTGGGTTTTCACTGAGCACTTTTGGTATGGGTGTAGGCTGGATTCGTCAGCCTTCAGGGAAGGGTCTGGAGTGGCTGGCACACATTTGGTGGGATGATGATAAGTACTATAACCCAGCCCTGAAGAGTCGGCTCACAATCTCCAAGGATACCTCCAAAAACCAGGTATTCCTCAAGATCGCCAATGTGGACACTGCAGATACTGCCACATACTACTGTGCTCGACCCCCTCTAACTGGTTTTGCTTACTGGGGCCAAGGGACTCTGGTCACTGTCTCTGCAGAGAGTCAGTCCTTCCCAAATGTCTTCCCCCTCGTCTCCTGCGAGAGCCCCCTGTCTGATAAGAATCTGGTGGCCATGGGCTGCCTGGCCCGGGACTTCCTGCCCAGCACCATTTCCTTCACCTGGAACTACCAGAACAACACTGAAGTCATCCAGGGTATCAGAACCTTCCCAACACTGAGGACAGGGGGCAAGTACCTAGCCACCTCGCA
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                VJ_sequence_nt_raw
#> 1                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                      ACTGATCAGTCTCCTCAGGCTGTCTCCTCAGGTTGCCTCCTCAAAATGAAGTTGCCTGTTAGGCTGTTGGTGCTGATGTTCTGGATTCCTGCTTCCAGCAGTGATGTTTTGATGACCCAAACTCCACTCTCCCTGCCTGTCAGTCTTGGAGATCAAGCCTCCATCTCTTGCAGATCTAGTCAGAGCATTGTACATAGTAATGGAAACACCTATTTAGAATGGTACCTGCAGAAACCAGGCCAGTCTCCAAAGCTCCTGATCTACAAAGTTTCCAACCGATTTTCTGGGGTCCCAGACAGGTTCAGTGGCAGTGGATCAGGGACAGATTTCACACTCAAGATCAGCAGAGTGGAGGCTGAGGATCTGGGAGTTTATTACTGCTTTCAAGGTTCACATGTTCCGTGGACGTTCGGTGGAGGCACCAAGCTGGAAATCAAACGGGCTGATGCTGCACCAACTGTATCCATCTTCCCACCATCCAGTGAGCAGTTAACATCTGGAGGTGCCTCAGTCGTGTGCTTC
#> 2                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                
#> 3                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                            GGGAGGAGACGTTGTAGAAATGAGACCGTCTATTCAGTTCCTGGGGCTCTTGTTGTTCTGGCTTCATGGTGCTCAGTGTGACATCCAGATGACACAGTCTCCATCCTCACTGTCTGCATCTCTGGGAGGCAAAGTCACCATCACTTGCAAGGCAAGCCAAGACATTAACAAGTATATAGCTTGGTACCAACACAAGCCTGGAAAAGGTCCTAGGCTGCTCATACATTACACATCTACATTACAGCCAGGCATCCCATCAAGGTTCAGTGGAAGTGGGTCTGGGAGAGATTATTCCTTCAGCATCAGCAACCTGGAGCCTGAAGATATTGCAACTTATTATTGTCTACAGTATGATAATCTTCTGTGGACGTTCGGTGGAGGCACCAAGCTGGAAATCAAACGGGCTGATGCTGCACCAACTGTATCCATCTTCCCACCATCCAGTGAGCAGTTAACATCTGGAGGTGCCTCAGTCGTGTGCTTC
#> 4                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                 GACTTTTGACTCACCATATCAAGTTCGCAGAATGAGGTTCTCTGCTCAGCTTCTGGGGCTGCTTGTGCTCTGGATCCCTGGATCCACTGCAGATATTGTGATGACGCAGGCTGCATTCTCCAATCCAGTCACTCTTGGAACATCAGCTTCCATCTCCTGCAGGTCTAGTAAGAGTCTCCTACATAGTAATGGCATCACTTATTTGTATTGGTATCTGCAGAAGCCAGGCCAGTCTCCTCAGCTCCTGATTTATCAGATGTCCAACCTTGCCTCAGGAGTCCCAGACAGGTTCAGTAGCAGTGGGTCAGGAACTGATTTCACACTGAGAATCAGCAGAATGGAGGCTGAGGATGTGGGTGTTTATTACTGTGCTCAAAATCTAGAACTTCCGTGGACGTTCGGTGGAGGCACCAAGCTGGAAATCAAACGGGCTGATGCTGCACCAACTGTATCCATCTTCCCACCATCCAGTGAGCAGTTAACATCTGGAGGTGCCTCAGTCGTGTGCTTC
#> 5                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             TGGGGACTTTTGACTCACCATATCAAGTTCGCAGAATGAGGTTCTCTGCTCAGCTTCTGGGGCTGCTTGTGCTCTGGATCCCTGGATCCACTGCAGATATTGTGATGACGCAGGCTGCATTCTCCAATCCAGTCACTCTTGGAACATCAGCTTCCATCTCCTGCAGGTCTAGTAAGAGTCTCCTACATACTAATGGCATCACTTATTTGTATTGGTATCTGCAGAAGCCAGGCCAGTCTCCTCAGCTCCTGATTTATCAGATGTCCAACCTTGCCTCAGGAGTCCCAGACAGGTTCAGTAGCAGTGGGTCAGGAACTGATTTCACACTGAGAATCAGCAGAGTGGAGGCTGAGGATGTGGGTGTTTATTACTGTGCTCAAAATCTTGAACTTCCGTACACGTTCGGAGGGGGGACCAAGCTGGAAATAAAACGGGCTGATGCTGCACCAACTGTATCCATCTTCCCACCATCCAGTGAGCAGTTAACATCTGGAGGTGCCTCAGTCGTGTGCTTC
#> 6  TTCTATGGGGATGGTCCACACAAACTCAGGGAAAGTTTGAAGATGGTATCCACACCTCAGTTCCTTGTATTTTTGCTTTTCTGGATTCCAGCCTCCAGAGGTGACATCTTGCTGACTCAGTCTCCAGCCATCCTGTCTGTGAGTCCAGGAGAAAGAGTCAGTTTCTCCTGCAGGGCCAGTCAGAGCATTGGCACAAGCATACACTGGTATCAGCAAAGAACAAATGGTTCTCCAAGGCTTCTCATAAAGTATGCTTCTGAGTCTATCTCTGGGATCCCTTCCAGGTTTAGTGGCAGTGGATCAGGGACAGATTTTACTCTTAGCATCAACAGTGTGGAGTCTGAAGATATTGCAGATTATTACTGTCAACAAAGTAATAGCTGGCCGCTCACGTTCGGTGCTGGGACCAAGCTGGAGCTGAAACGGGCTGATGCTGCACCAACTGTATCCATCTTCCCACCATCCAGTGAGCAGTTAACATCTGGAGGTGCCTCAGTCGTGTGCTTC;TGGGGGACTGAGATGGAAAACAAAATGGATTTTCAGATGCAGATTATCAGCTTGCTGCTAATCAGTGTCACAGTCATAGTGTCTAATGGAGAAATTGTGCTCACCCAGTCTCCAACCACCATGGCTGCATCTCCCGGGGAGAAGATCACTATCACCTGCAGTGCCAGCTCAAGTATAAGTTCCAATTACTTGCATTGGTATCAGCAGAAGCCAGGATTCTCCCCTAAACTCTTGATTTATAGGACATCCAATCTGGCTTCTGGAGTCCCAGCTCGCTTCAGTGGCAGTGGGTCTGGGACCTCTTACTCTCTCACAATTGGCACCATGGAGGCTGAAGATGTTGCCACTTACTACTGCCAGCAGGGTAGTAGTATACCGCTCACGTTCGGTGCTGGGACCAAGCTGGAGCTGAAACGGGCTGATGCTGCACCAACTGTATCCATCTTCCCACCATCCAGTGAGCAGTTAACATCTGGAGGTGCCTCAGTCGTGTGCTTC
#> 7  TTCTATGGGGATGGTCCACACAAACTCAGGGAAAGTTTGAAGATGGTATCCACACCTCAGTTCCTTGTATTTTTGCTTTTCTGGATTCCAGCCTCCAGAGGTGACATCTTGCTGACTCAGTCTCCAGCCATCCTGTCTGTGAGTCCAGGAGAAAGAGTCAGTTTCTCCTGCAGGGCCAGTCAGAGCATTGGCACAAGCATACACTGGTATCAGCAAAGAACAAATGGTTCTCCAAGGCTTCTCATAAAGTATGCTTCTGAGTCTATCTCTGGGATCCCTTCCAGGTTTAGTGGCAGTGGATCAGGGACAGATTTTACTCTTAGCATCAACAGTGTGGAGTCTGAAGATATTGCAGATTATTACTGTCAACAAAGTAATAGCTGGCCGCTCACGTTCGGTGCTGGGACCAAGCTGGAGCTGAAACGGGCTGATGCTGCACCAACTGTATCCATCTTCCCACCATCCAGTGAGCAGTTAACATCTGGAGGTGCCTCAGTCGTGTGCTTC;TGGGGGACTGAGATGGAAAACAAAATGGATTTTCAGATGCAGATTATCAGCTTGCTGCTAATCAGTGTCACAGTCATAGTGTCTAATGGAGAAATTGTGCTCACCCAGTCTCCAACCACCATGGCTGCATCTCCCGGGGAGAAGATCACTATCACCTGCAGTGCCAGCTCAAGTATAAGTTCCAATTACTTGCATTGGTATCAGCAGAAGCCAGGATTCTCCCCTAAACTCTTGATTTATAGGACATCCAATCTGGCTTCTGGAGTCCCAGCTCGCTTCAGTGGCAGTGGGTCTGGGACCTCTTACTCTCTCACAATTGGCACCATGGAGGCTGAAGATGTTGCCACTTACTACTGCCAGCAGGGTAGTAGTATACCGCTCACGTTCGGTGCTGGGACCAAGCTGGAGCTGAAACGGGCTGATGCTGCACCAACTGTATCCATCTTCCCACCATCCAGTGAGCAGTTAACATCTGGAGGTGCCTCAGTCGTGTGCTTC
#> 8                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                   GGGGACCAATATTGAAAAGAATAGACCTGGTTTGTGAATTATGGCCTGGATTTCACTTATACTCTCTCTCCTGGCTCTCAGCTCAGGGGCCATTTCCCAGGCTGTTGTGACTCAGGAATCTGCACTCACCACATCACCTGGTGAAACAGTCACACTCACTTGTCGCTCAAGTACTGGGGCTGTTACAACTAGTAACTATGCCAACTGGGTCCAAGAAAAACCAGATCATTTATTCACTGGTCTAATAGGTGGTACCAACAACCGAGCTCCAGGTGTTCCTGCCAGATTCTCAGGCTCCCTGATTGGAGACAAGGCTGCCCTCACCATCACAGGGGCACAGACTGAGGATGAGGCAATATATTTCTGTGCTCTATGGTACAGCAACCATTTGGTGTTCGGTGGAGGAACCAAACTGACTGTCCTAGGCCAGCCCAAGTCTTCGCCATCAGTCACCCTGTTTCCACCTTCCTCTGAAGAGCTCGAGACTAACAAGGCCACACTGGTGTGTA
#> 9                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                              GGATTGTCATTGCAGCCAGGACTCAGCATGGACATGAGGACCCCTGCTCAGTTTCTTGGAATCTTGTTGCTCTGGTTTCCAGGTATCAAATGTGACATCAAGATGACCCAGTCTCCATCTTCCATGTATGCATCTCTAGGAGAGAGAGTCACTATCACTTGCAAGGCGAGTCAGGACATTAATAGCTATTTAAGCTGGTTCCAGCAGAAACCAGGGAAATCTCCTAAGACCCTGATCTATCGTGCAAACAGATTGGTAGATGGGGTCCCATCAAGGTTCAGTGGCAGTGGATCTGGGCAAGATTTTTCTCTCACCATCAGCAGCCTGGAGTATGAAGATATGGGAATTTATTATTGTCTACAGTATGATGAGTTTCCGTCCACGTTCGGAGGGGGGACCAACCTGGAAATAAAACGGGCTGATGCTGCACCAACTGTATCCATCTTCCCACCATCCAGTGAGCAGTTAACATCTGGAGGTGCCTCAGTCGTGTGCTTC
#> 10                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                       TTATGGGGATTGTCATTGCAGTCAGGACTCAGCATGGACATGAGGGCTCCTGCACAGATTTTTGGCTTCTTGTTGCTCTTGTTTCCAGGTACCAGATGTGACATCCAGATGACCCAGTCTCCATCCTCCTTATCTGCCTCTCTGGGAGAAAGAGTCAGTCTCACTTGTCGGGCAAGTCAGGACATTGGTAGTAGCTTAAACTGGCTTCAGCAGGAACCAGATGGAACTATTAAACGCCTGATCTACGCCACATCCAGTTTAGATTCTGGTGTCCCCAAAAGGTTCAGTGGCAGTAGGTCTGGGTCAGATTATTCTCTCACCATCAGCAGCCTTGAGTCTGAAGATTTTGTAGACTATTACTGTCTACAATATGCTAGTTCTCCGTGGACGTTCGGTGGAGGCACCAAGCTGGAAATCAAACGGGCTGATGCTGCACCAACTGTATCCATCTTCCCACCATCCAGTGAGCAGTTAACATCTGGAGGTGCCTCAGTCGTGTGCTTC
#> 11                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                          GGGGGATCTACATCTGAAAGGCAGGTGGAGCAAGATGGAATCACAGACTCAGGTCCTCATGTCCCTGCTGTTCTGGGTATCTGGTACCTGTGGGGACATTGTGATGACACAGTCTCCATCCTCCCTGACTGTGACAGCAGGAGAGAAGGTCACTATGAGCTGCAAGTCCAGTCAGAGTCTGTTAAACAGTGGAAATCAAAAGAACTACTTGACCTGGTACCAGCAGAAACCAGGGCAGCCTCCTAAACTGTTGATCTACTGGGCATCCACTAGGGAATCTGGGGTCCCTGATCGCTTCACAGGCAGTGGATCTGGAACAGATTTCACTCTCACCATCAGCAGTGTGCAGGCTGAAGACCTGGCAGTTTATTACTGTCAGAATGATTATAGTTATCCGCTCACGTTCGGTGCTGGGACCAAGCTGGAGCTGAAACGGGCTGATGCTGCACCAACTGTATCCATCTTCCCACCATCCAGTGAGCAGTTAACATCTGGAGGTGCCTCAGTCGTGTGCTTC
#> 12                                                                                                                                                                                                                                                                                                                                                                                                                                                GAGAACTACAACCTGTCTGTCTCAGCAGAGATCAGTAGTACCTGCATTATGGCCTGGACTCCTCTCTTCTTCTTCTTTGTTCTTCATTGCTCAGGTTCTTTCTCCCAACTTGTGCTCACTCAGTCATCTTCAGCCTCTTTCTCCCTGGGAGCCTCAGCAAAACTCACGTGCACCTTGAGTAGTCAGCACAGTACGTACACCATTGAATGGTATCAGCAACAGCCACTCAAGCCTCCTAAGTATGTGATGGAGCTTAAGAAAGATGGAAGCCACAGCACAGGTGATGGGATTCCTGATCGCTTCTCTGGATCCAGCTCTGGTGCTGATCGCTACCTTAGCATTTCCAACATCCAGCCTGAAGATGAAGCAATATACATCTGTGGTGTGGGTGATACAATTAAGGAACAATTTGTGTATGTTTTCGGCGGTGGAACCAAGGTCACTGTCCTAGGTCAGCCCAAGTCCACTCCCACTCTCACCGTGTTTCCACCTTCCTCTGAGGAGCTCAAGGAAAACAAAGCCACACTGGTGTGTCTGATTTCCAACTTTTCCCCGAGTGGTGTGACAGTGGCCTG
#> 13                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                GAGTCAGCCTCACACTGATCACACACAGACATGAGTGTGCCCACTCAGGTCCTGGGGTTGCTGCTGCTGTGGCTTACAGATGCCAGATGTGACATCCAGATGACTCAGTCTCCAGCCTCCCTATCTGTATCTGTGGGAGAAACTGTCACCATCACATGTCGAGCAAGTGAGAATATTTACAGTAATTTAGCATGGTATCAGCAGAAACAGGGAAAATCTCCTCAGCTCCTGGTCTATGCTGCAACAAACTTAGCAGATGGTGTGCCATCAAGGTTCAGTGGCAGTGGATCAGGCACACAGTATTCCCTCAAGATCAACAGCCTGCAGTCTGAAGATTTTGGGAGTTATTACTGTCAACATTTTTGGGGTACTCCTCGGACGTTCGGTGGAGGCACCAAGCTGGAAATCAAACGGGCTGATGCTGCACCAACTGTATCCATCTTCCCACCATCCAGTGAGCAGTTAACATCTGGAGGTGCCTCAGTCGTGTGCTTC
#> 14                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                              GGGAAATACATCAGATCAGCATGGGCATCAAGATGGAGTCACAGACTCAGGTCTTTGTATACATGTTGCTGTGGTTGTCTGGTGTTGATGGAGACATTGTGATGACCCAGTCTCAAAAATTCATGTCCACATCAGTAGGAGACAGGGTCAGCGTCACCTGCAAGGCCAGTCAGAATGTGGGTACTAATGTAGCCTGGTATCAACAGAAACCAGGGCAATCTCCTAAAGCACTGATTTACTCGGCATCCTACCGGTACAGTGGAGTCCCTGATCGCTTCACAGGCAGTGGATCTGGGACAGATTTCACTCTCACCATCAGCAATGTGCAGTCTGAAGACTTGGCAGAGTATTTCTGTCAGCAATATAACAGCTATCCGCTCACGTTCGGTGCTGGGACCAAGCTGGAGCTGAAACGGGCTGATGCTGCACCAACTGTATCCATCTTCCCACCATCCAGTGAGCAGTTAACATCTGGAGGTGCCTCAGTCGTGTGCTTC
#> 15                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                GAAATGCATCACACCAGCATGGGCATCAAAATGGAGTCACAGATTCAGGTCTTTGTATTCGTGTTTCTCTGGTTGTCTGGTGTTGACGGAGACATTGTGATGACCCAGTCTCACAAATTCATGTCCACATCAGTAGGAGACAGGGTCAGCATCACCTGCAAGGCCAGTCAGGATGTGAGTACTGCTGTAGCCTGGTATCAACAGAAACCAGGACAATCTCCTAAACTACTGATTTACTCGGCATCCTACCGGTACACTGGAGTCCCTGATCGCTTCACTGGCAGTGGATCTGGGACGGATTTCACTTTCACCATCAGCAGTGTGCAGGCTGAAGACCTGGCAGTTTATTACTGTCAGCAACATTATAGTACTCCATTCACGTTCGGCTCGGGGACAAAGTTGGAAATAAAACGGGCTGATGCTGCACCAACTGTATCCATCTTCCCACCATCCAGTGAGCAGTTAACATCTGGAGGTGCCTCAGTCGTGTGCTTC
#> 16                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                               GGGGAGTCATTCTTGGTCAGGAGACGTTGTAGAAATGAGACCGTCTATTCAGTTCCTGGGGCTCTTGTTGTTCTGGCTTCATGGTGCTCAGTGTGACATCCAGATGACACAGTCTCCATCCTCACTGTCTGCATCTCTGGGAGGCAAAGTCACCATCACTTGCAAGGCAAGCCAAGACATTAACAAGTATATAGCTTGGTACCAACACAAGCCTGGAAAAGGTCCTAGGCTGCTCATACATTACACATCTACATTACAGCCAGGCATCCCATCAAGGTTCAGTGGAAGTGGGTCTGGGAGAGATTATTCCTTCAGCATCAGCAACCTGGAGCCTGAAGATATTGCAACTTATTATTGTCTACAGTATGATAATCTATTCACGTTCGGCTCGGGGACAAAGTTGGAAATAAAACGGGCTGATGCTGCACCAACTGTATCCATCTTCCCACCATCCAGTGAGCAGTTAACATCTGGAGGTGCCTCAGTCGTGTGCTTC
#> 17                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                           ATCCTCTCTTCCAGCTCTCAGAGATGGAGACAGACACACTCCTGCTATGGGTGCTGCTGCTCTGGGTTCCAGGTTCCACAGGTGACATTGTGCTGACCCAATCTCCAGCTTCTTTGGCTGTGTCTCTAGGGCAGAGGGCCACCATATCCTGCAGAGCCAGTGAAAGTGTTGATAGTTATGGCAATAGTTTTATGCACTGGTACCAGCAGAAACCAGGACAGCCACCCAAACTCCTCATCTATCGTGCATCCAACCTAGAATCTGGGATCCCTGCCAGGTTCAGTGGCAGTGGGTCTAGGACAGACTTCACCCTCACCATTAATCCTGTGGAGGCTGATGATGTTGCAACCTATTACTGTCAGCAAAGTAATGAGGATCCTCGGACGTTCGGTGGAGGCACCAAGCTGGAAATCAAACGGGCTGATGCTGCACCAACTGTATCCATCTTCCCACCATCCAGTGAGCAGTTAACATCTGGAGGTGCCTCAGTCGTGTGCTTC
#> 18                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                GAAATGCATCACACCAGCATGGGCATCAAAATGGAGTCACAGATTCAGGTCTTTGTATTCGTGTTTCTCTGGTTGTCTGGTGTTGACGGAGACATTGTGATGACCCAGTCTCACAAATTCATGTCCACATCAGTAGGAGACAGGGTCAGCATCACCTGCAAGGCCAGTCAGGATGTGAGTACTGCTGTAGCCTGGTATCAACAGAAACCAGGACAATCTCCTAAACTACTGATTTACTCGGCATCCTACCGGTACACTGGAGTCCCTGATCGCTTCACTGGCAGTGGATCTGGGACGGATTTCACTTTCACCATCAGCAGTGTGCAGGCTGAAGACCTGGCAGTTTATTACTGTCAGCAACATTATAGTACTCCGTACACGTTCGGAGGGGGGACCAAGCTGGAAATAAAACGGGCTGATGCTGCACCAACTGTATCCATCTTCCCACCATCCAGTGAGCAGTTAACATCTGGAGGTGCCTCAGTCGTGTGCTTC
#> 19                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                            ATTGGGGTCTGCATCAGAAAGGCAGGGGGATCAAGATGGAATCACAGACTCAGGTCTTCCTCTCCCTGCTGCTCTGGGTATCTGGTACCTGTGGGAACATTATGATGACACAGTCGCCATCATCTCTGGCTGTGTCTGCAGGAGAAAAGGTCACTATGAGCTGTAAGTCCAGTCAAAGTGTTTTATACAGTTCAAATCAGAAGAACTACTTGGCCTGGTACCAGCAGAAACCAGGGCAGTCTCCTAAACTGCTGATCTACTGGGCATCCACTAGGGAATCTGGTGTCCCTGATCGCTTCACAGGCAGTGGATCTGGGACAGATTTTACTCTTACCATCAGCAGTGTACAAGCTGAAGACCTGGCAGTTTATTACTGTCATCAATACCTCTCCTCGTGGACGTTCGGTGGAGGCACCAAGCTGGAAATCAAACGGGCTGATGCTGCACCAACTGTATCCATCTTCCCACCATCCAGTGAGCAGTTAACATCTGGAGGTGCCTCAGTCGTGTGCTTC
#>    VDJ_sequence_nt_trimmed VJ_sequence_nt_trimmed VDJ_sequence_aa
#> 1                       NA                                     NA
#> 2                       NA                                     NA
#> 3                       NA                                     NA
#> 4                       NA                                     NA
#> 5                       NA                                     NA
#> 6                       NA                                     NA
#> 7                       NA                                     NA
#> 8                       NA                                     NA
#> 9                       NA                                     NA
#> 10                      NA                                     NA
#> 11                      NA                                     NA
#> 12                      NA                                     NA
#> 13                      NA                                     NA
#> 14                      NA                                     NA
#> 15                      NA                                     NA
#> 16                      NA                                     NA
#> 17                      NA                                     NA
#> 18                      NA                                     NA
#> 19                      NA                                     NA
#>    VJ_sequence_aa VDJ_trimmed_ref VJ_trimmed_ref       VDJ_raw_consensus_id
#> 1                              NA                                          
#> 2                              NA                clonotype2053_concat_ref_1
#> 3                              NA                clonotype2386_concat_ref_1
#> 4                              NA                                          
#> 5                              NA                clonotype2040_concat_ref_1
#> 6                              NA                 clonotype516_concat_ref_2
#> 7                              NA                 clonotype516_concat_ref_2
#> 8                              NA                 clonotype141_concat_ref_1
#> 9                              NA                                          
#> 10                             NA                 clonotype390_concat_ref_1
#> 11                             NA                 clonotype464_concat_ref_1
#> 12                             NA                                          
#> 13                             NA                clonotype1725_concat_ref_1
#> 14                             NA                clonotype1243_concat_ref_1
#> 15                             NA                clonotype2271_concat_ref_1
#> 16                             NA                clonotype1028_concat_ref_1
#> 17                             NA                 clonotype592_concat_ref_1
#> 18                             NA                 clonotype649_concat_ref_1
#> 19                             NA                 clonotype634_concat_ref_1
#>           VJ_raw_consensus_id     orig_barcode clonotype_frequency specifity
#> 1     clonotype7_concat_ref_1 TATGCCCCAGCTTCGG                  25        NA
#> 2                             CCCAGTTAGCAGGTCA                   1        NA
#> 3  clonotype2386_concat_ref_2 CGTTGGGAGGGATCTG                   1        NA
#> 4   clonotype441_concat_ref_1 CGAACATAGTGACATA                   1        NA
#> 5  clonotype2040_concat_ref_2 ACGATGTTCAAACCGT                   1        NA
#> 6   clonotype516_concat_ref_4 AGGTCATTCTTGGGTA                   1        NA
#> 7   clonotype516_concat_ref_4 AGGTCATTCTTGGGTA                   1        NA
#> 8   clonotype141_concat_ref_2 ATGCGATTCCCATTAT                   2        NA
#> 9   clonotype743_concat_ref_1 ATGAGGGCAATGAATG                   1        NA
#> 10  clonotype390_concat_ref_2 AGCGTATCACGGACAA                   1        NA
#> 11  clonotype464_concat_ref_2 GTGCGGTAGCCCAGCT                   1        NA
#> 12   clonotype10_concat_ref_1 TTCTTAGCATTAGGCT                   6        NA
#> 13 clonotype1725_concat_ref_2 GCTGGGTGTCCAGTTA                   1        NA
#> 14 clonotype1243_concat_ref_2 CTACGTCCAAAGGCGT                   1        NA
#> 15 clonotype2271_concat_ref_2 GATCGATAGTTAGCGG                   1        NA
#> 16 clonotype1028_concat_ref_2 TGGGCGTTCGGTTCGG                   1        NA
#> 17  clonotype592_concat_ref_2 ATTATCCAGTTCGATC                   1        NA
#> 18  clonotype649_concat_ref_2 CCTCTGACACCAGATT                   1        NA
#> 19  clonotype634_concat_ref_2 CTACCCAAGACTAGAT                   1        NA
#>    affinity GEX_available    orig.ident orig_barcode_GEX seurat_clusters
#> 1        NA         FALSE          <NA>             <NA>            <NA>
#> 2        NA         FALSE          <NA>             <NA>            <NA>
#> 3        NA         FALSE          <NA>             <NA>            <NA>
#> 4        NA         FALSE          <NA>             <NA>            <NA>
#> 5        NA         FALSE          <NA>             <NA>            <NA>
#> 6        NA          TRUE SeuratProject AGGTCATTCTTGGGTA               0
#> 7        NA          TRUE SeuratProject AGGTCATTCTTGGGTA               0
#> 8        NA          TRUE SeuratProject ATGCGATTCCCATTAT               0
#> 9        NA         FALSE          <NA>             <NA>            <NA>
#> 10       NA         FALSE          <NA>             <NA>            <NA>
#> 11       NA          TRUE SeuratProject GTGCGGTAGCCCAGCT               1
#> 12       NA         FALSE          <NA>             <NA>            <NA>
#> 13       NA         FALSE          <NA>             <NA>            <NA>
#> 14       NA         FALSE          <NA>             <NA>            <NA>
#> 15       NA          TRUE SeuratProject GATCGATAGTTAGCGG               0
#> 16       NA          TRUE SeuratProject TGGGCGTTCGGTTCGG               1
#> 17       NA          TRUE SeuratProject ATTATCCAGTTCGATC               2
#> 18       NA          TRUE SeuratProject CCTCTGACACCAGATT               4
#> 19       NA          TRUE SeuratProject CTACCCAAGACTAGAT               2
#>          PC_1      PC_2    UMAP_1      UMAP_2     tSNE_1     tSNE_2     batches
#> 1          NA        NA        NA          NA         NA         NA Unspecified
#> 2          NA        NA        NA          NA         NA         NA Unspecified
#> 3          NA        NA        NA          NA         NA         NA Unspecified
#> 4          NA        NA        NA          NA         NA         NA Unspecified
#> 5          NA        NA        NA          NA         NA         NA Unspecified
#> 6  -0.9356794 -5.038539 -7.042369 -1.96977423 -16.495634  22.838189 Unspecified
#> 7  -0.9356794 -5.038539 -7.042369 -1.96977423 -16.495634  22.838189 Unspecified
#> 8  -2.4723771 -0.819026 -2.945280 -4.05283188 -13.722256  -1.487648 Unspecified
#> 9          NA        NA        NA          NA         NA         NA Unspecified
#> 10         NA        NA        NA          NA         NA         NA Unspecified
#> 11 -2.3438558  8.659659 12.260447 -1.46314500  -3.783243 -29.142588 Unspecified
#> 12         NA        NA        NA          NA         NA         NA Unspecified
#> 13         NA        NA        NA          NA         NA         NA Unspecified
#> 14         NA        NA        NA          NA         NA         NA Unspecified
#> 15 -2.4256903 -4.098689 -3.537320 -2.80273508 -20.631126   6.718939 Unspecified
#> 16 -3.7693754  9.494090 11.363246 -2.93161844  -8.847022 -27.030247 Unspecified
#> 17  0.1115660 -2.332551 -6.830872  3.64487959  19.729280  10.202692 Unspecified
#> 18 -2.3275297 -3.097667 -5.545405 -0.09513162  -3.070754  20.088388 Unspecified
#> 19 -0.6573374 -6.195999 -6.550993  1.56593950   9.316681  11.732505 Unspecified
#>     clonotype_id   FB_assignment
#> 1     clonotype7 Dummy_barcode_1
#> 2  clonotype2053 Dummy_barcode_1
#> 3  clonotype2386 Dummy_barcode_1
#> 4   clonotype441 Dummy_barcode_1
#> 5  clonotype2040 Dummy_barcode_1
#> 6   clonotype516 Dummy_barcode_1
#> 7   clonotype516 Dummy_barcode_1
#> 8   clonotype141 Dummy_barcode_1
#> 9   clonotype743 Dummy_barcode_1
#> 10  clonotype390 Dummy_barcode_1
#> 11  clonotype464 Dummy_barcode_1
#> 12   clonotype10 Dummy_barcode_1
#> 13 clonotype1725 Dummy_barcode_1
#> 14 clonotype1243 Dummy_barcode_1
#> 15 clonotype2271 Dummy_barcode_1
#> 16 clonotype1028 Dummy_barcode_1
#> 17  clonotype592 Dummy_barcode_1
#> 18  clonotype649 Dummy_barcode_1
#> 19  clonotype634 Dummy_barcode_1
#>                                                     CDR3aa expanded_number  X
#> 1                                              CFQGSHVPWTF              NA  1
#> 2                                         CARSSLYYGNFYFDYW              NA  2
#> 3                          CTRNMRLRRGTGTGYAMDYWCLQYDNLLWTF              NA  3
#> 4                                              CAQNLELPWTF              NA  4
#> 5                              CVRGGYDYDRDYFDFWCAQNLELPYTF              NA  5
#> 6  CARNRITTVVAPMDYWCQQSNSWPLTF;CARNRITTVVAPMDYWCQQGSSIPLTF               2  6
#> 7          CARKDYARGDYWCQQSNSWPLTF;CARKDYARGDYWCQQGSSIPLTF               2  7
#> 8                            CARWGIYDGYYGDAMDYWCALWYSNHLVF              NA  8
#> 9                                              CLQYDEFPSTF              NA  9
#> 10                                     CARTFAYWCLQYASSPWTF              NA 10
#> 11                                   CARMVTGAYWCQNDYSYPLTF              NA 11
#> 12                                         CGVGDTIKEQFVYVF              NA 12
#> 13                                 CARVNLHAMDYWCQHFWGTPRTF              NA 13
#> 14                                     CVRSWDYWCQQYNSYPLTF              NA 14
#> 15                           CTFRIYYYGSSPYYFDYWCQQHYSTPFTF              NA 15
#> 16                             CAREGYYGSSPYAMDYWCLQYDNLFTF              NA 16
#> 17                                 CAREATVVADYWCQQSNEDPRTF              NA 17
#> 18                               CASGGGNPWYFDVWCQQHYSTPYTF              NA 18
#> 19                                  CARPPLTGFAYWCHQYLSSWTF              NA 19
#>  [ reached 'max' / getOption("max.print") -- omitted 32 rows ]