Function to get shared/public elements across multiple repertoires

Function to get shared elements across multiple repertoires, specified by the feature.columns parameter (a column of the VDJ matrix). If two columns are specified in feature.columns, the resulting shared features will combine the values from each column (at a per-cell level).

VDJ_get_public(
  VDJ,
  feature.columns,
  repertoire.column,
  specific.repertoires,
  find.public.all,
  find.public.percentage,
  treat.combined.features,
  output.format
)

Arguments

VDJ
VDJ or VDJ.GEX.matrix[[1]] object, as obtained from the VDJ_GEX_matrix function in Platypus.

feature.columns
Character or character vector columns of features to be assayed

repertoire.column
string - the repertoire-defining column (default to 'sample_id').

specific.repertoires
vector of strings or NULL - if only the shared elements from specific repertoires should be taken into account. If NULL, will output the shared/public elements across all repertoires.

find.public.all
boolean - if T, will look for the public elements across all repertoires

find.public.percentage
list - the first element denotes the percentage of repertoires to get shared elements for, the second element is the maximum number of repertoire combinations to consider (can be NULL to consider all).

treat.combined.features
string - 'exclude' will exclude combined features with one element missing, 'include' will include and considers them as a new feature value.

output.format
string - 'df' to get a shared element dataframe (with columns = Repertoire and Public), 'list' for a list of shared elements.

Value

Either a dataframe of public elements across multiple repertoires or a list.

Examples

VDJ_get_public(VDJ = small_vgm[[1]],
feature.columns='VDJ_cdr3s_aa', find.public.all=TRUE,
output.format='df')
#> [1] Repertoires Public     
#> <0 Zeilen> (oder row.names mit Länge 0)