Yields overlap heatmap and datatable of features or combined features for different samples or groups

Yields overlap heatmap and datatable of features or combined features for different samples or groups

VDJ_overlap_heatmap(
  VDJ,
  feature.columns,
  grouping.column,
  jaccard,
  plot.type,
  pvalues.label.size,
  axis.label.size,
  add.barcode.table
)

Arguments

VDJ
VDJ output of the VDJ_GEX_matrix function (VDJ_GEX_matrix.output[[1]])

feature.columns
A character array of column names of which the overlap should be displayed. The content of these columns is pasted together (separated by "/"). E.g. if the overlap in cells germline gene usage is desired, the input could be c("VDJ_jgene","VDJ_dgene","VDJ_vgene"). These columns would be pasted and compared across the grouping variable.

grouping.column
A column which acts as a grouping variable. If repertoires are to be compared use the sample_id column.

jaccard
Boolean. Defaults to FALSE. If set to TRUE, the overlap will be reported as jaccard index. If set to FALSE the overlap will be repored as absolute counts

plot.type
Character. Either "ggplot" or "pheatmap". Defaults to Pheatmap

pvalues.label.size
Numeric. Defaults to 4. Is passed on to ggplot theme

axis.label.size
Numeric. Defaults to 4. Is passed on to ggplot theme

add.barcode.table
Boolean. Defaults to T. Whether to generate a dataframe with frequencies and barcodes of cells with overlapping features. This is useful to e.g. analyze deferentially expressed genes between cells of two samples or groups expressing the same VDJ or VJ chain

Value

A list of a ggplot (out[[1]]), the source table or matrix for the plot out[[2]] and a table containing additional information in case that add.barcode.table was set to TRUE (out[[3]])

Examples

#To test the overlap of CDR3s between multiple samples
overlap <- VDJ_overlap_heatmap(VDJ = Platypus::small_vgm[[1]]
,feature.columns = c("VDJ_cdr3s_aa"),
grouping.column = "sample_id", axis.label.size = 15
, plot.type = "ggplot")