Produces neighbor joining phylogenetic trees from the output of VDJ_clonal_lineages

Produces neighbor joining phylogenetic trees from the output of VDJ_clonal_lineages

VDJ_tree(
  clonal.lineages,
  with.germline,
  min.sequences,
  max.sequences,
  normalize.germline.length,
  unique.sequences
)

Arguments

clonal.lineages
Output from VDJ_clonal_lineages. This should be nested list, with the outer list element corresponding to the individual repertoire and the inner list corresponding to individual clonal lineages based on the initial clonotyping strategy in the form of a dataframe with either Seq or Name.

with.germline
Logical specifying if the germline should be set as outgroup. Default is set to TRUE.

min.sequences
integer value specifying the minimum number of sequences to be allowed for clonal lineages. Default is 3.

max.sequences
integer value specifying the maximum number of sequences to be allowed for clonal lineages. Default is 500

normalize.germline.length
Logical determining whether or not the branch length separating the germline from the first internal node should be normalized. Potentially useful for visualization if the remainder tips are far from the root. Default is TRUE.

unique.sequences
Logical indicating if those cells containing identical VDJRegion sequences should be merged into single nodes and have their variant added as the tip label. Default is TRUE.

Value

Returns a nested list of phylogenetic trees. The output[[i]][[j]] corresponds to the j'th clone in the i'th input repertoire. plot(output[[i]][[j]]) should display the phylogenetic tree if the ape package is loaded.

Examples

if (FALSE) {
vdj.tree <- VDJ_tree(clonal.lineages = VDJ.clonal.lineage.output
,with.germline=T,min.sequences = 5
,max.sequences = 30,unique.sequences = T)
}