Get information about somatic hypermutation in the simulation. This function return a barplot showing the average mutation.

Get information about somatic hypermutation in the simulation. This function return a barplot showing the average mutation.

get.avr.mut.data(igraph.index.attr, history, clonotype.select, level)

Arguments

igraph.index.attr
A list "igraph.index.attr" from the simulation output.

history
A dataframe "history" from the simulation output.

clonotype.select
The selected clonotype index, can be the output of the function "select.top.clone.clone".

level
Can be "clone" or "cell". If "clone", the function will return average mutation on unique variant level. Otherwise it will return on cell level.

Value

a bar plot showing the average mutation on clone or cell level.