mus_t_trans A data frame contains mouse B cell average gene expression for multiple cell types, with the rows representing the gene names, column names representing the cell type names. The original single cell sequencing data is retrieved from 10xgenomics and combined with experimental data from.#? The expression level for different cell types are obtained by calculating the average expression after sorting the original data by markers as shown below.

NaiveCd4

Cd4+;Cd44-

MemoryCd4

Cd4+;Il7R+;Cd44+

ActivatedCd4

Cd4+;Cd44+;Il7R-;Ifng+

NaiveCd8

Cd8a+;Tcf7+;Cd44-

EffectorCd8

Cd8a+;Cx3cr1+;Il7R-

MemoryCd8

Cd8a+;Il7R+;Cd44+

ExhaustedCd8

Pdcd1+;Cd8a+

mus_t_trans A data frame contains mouse B cell average gene expression for multiple cell types, with the rows representing the gene names, column names representing the cell type names. The original single cell sequencing data is retrieved from 10xgenomics and combined with experimental data from.#? The expression level for different cell types are obtained by calculating the average expression after sorting the original data by markers as shown below.

NaiveCd4

Cd4+;Cd44-

MemoryCd4

Cd4+;Il7R+;Cd44+

ActivatedCd4

Cd4+;Cd44+;Il7R-;Ifng+

NaiveCd8

Cd8a+;Tcf7+;Cd44-

EffectorCd8

Cd8a+;Cx3cr1+;Il7R-

MemoryCd8

Cd8a+;Il7R+;Cd44+

ExhaustedCd8

Pdcd1+;Cd8a+

mus_t_trans

Format

A data frame with 26538 rows and 7 variables, with the rows representing the gene names, column names representing the cell type names.

Source

https://support.10xgenomics.com/single-cell-vdj/datasets https://support.10xgenomics.com/single-cell-vdj/datasets/3.0.0/vdj_v1_mm_c57bl6_pbmc_5gex https://support.10xgenomics.com/single-cell-vdj/datasets/3.0.0/vdj_v1_mm_balbc_pbmc_5gex