Simulate B or T cell receptor sequences by variational autoencodes(VAEs) trained with experimental data.

Simulate B or T cell receptor sequences by variational autoencodes(VAEs) trained with experimental data.

Echidna_vae_generate(
  sequence,
  n.train,
  n.sample,
  batch.size,
  latent.dim,
  intermediate.dim,
  epochs,
  epsilon.std,
  null.threshold
)

Arguments

sequence
a vector of seuqnece the model to be trained on

n.train
number of sequence to be used in training set, the rest will be in testing set

n.sample
number of new sequence to generate from VAE model

batch.size
set to larger to save time, set to smaller to same computing power

latent.dim
parameter used in VAE model

intermediate.dim
parameter used in VAE model

epochs
parameter used in VAE model

epsilon.std
parameter used in VAE model

null.threshold
threshold of predicted value to be considered as an existing base, default is 0.05. When generated sequence is too short, lower this threshold.