Loads and saves RData objects from the PlatypusDB

Loads and saves RData objects from the PlatypusDB

PlatypusDB_fetch(
  PlatypusDB.links,
  save.to.disk,
  load.to.enviroment,
  load.to.list,
  path.to.save,
  combine.objects
)

Arguments

PlatypusDB.links

Character vector. One or more links to files in the PlatypusDB. Links are constructed as follows: "%Project id%/%sample_id%/%filetype%". Any of the three can be "ALL", to download all files fitting the other link elements. If %filetype% is gexVGM, vdjVGM or metadata, %sample_id% needs to be "ALL", as these are elements which are not divided by sample. See examples for clarification. See last example on how to download AIRR compliant data. Feature Barcode (FB) data will be downloaded both for GEX and VDJ if present and does not need to be specified in the path

save.to.disk

Boolean. Defaults to FALSE. Whether to save downloaded files individually to the directory specified in path.to.save

load.to.enviroment

Boolean. Defaults to TRUE. Whether to load objects directly into the current .GlobalEnv. An array of the names of the loaded objects will be returned. !Be aware of RAM limitations of your machine when downloading multiple large files.

load.to.list

Boolean. Defaults to FALSE. Whether to return loaded objects as a list. !Be aware of RAM limitations of your machine when downloading multiple large files.

path.to.save

System path to save files to.

combine.objects

Boolean. Defaults to TRUE. Whether to combine objects if appropriate. e.g. VDJ and GEX RData objects for a sample are saved as two independent objects and downloaded as such, to allow for flexibility. If combine.objects is set to TRUE, the function will coerce RData objects of each loaded sample or of each loaded VDJ_GEX_matrix appropriately. Combined input of VDJ and GEX Rdata objects can be directly supplied to the VDJ_GEX_matrix function.

Value

A list of loaded project files as R objects if load.to.list = T or a name of these object loaded to the enviroment if load.to.enviroment = T.

Examples

# \donttest{

#Get a list of available projects by name
names(PlatypusDB_list_projects())
#>  [1] "agrafiotis2021a"     "horns2020a"          "khatun2019a"        
#>  [4] "kreiner2021a"        "kreiner2021b"        "kuhn2021a"          
#>  [7] "mathew2021a"         "merkenschlager2021a" "neumeier2021a"      
#> [10] "neumeier2021b"       "sandu2021a"          "shlesinger2022a"    
#> [13] "shlesinger2022b"     "shlesinger2022c"     "yermanos2021a"      
#> [16] "yermanos2021b"       "yermanos2022a"       "yermanos2022b"      
#> [19] "yermanos2022c"       "yewdell2021a"        "yost2019"           

#Load the VDJ_GEX_matrix of a project as an object and
#also save it to disk for later.
#This will download the VDJ and GEX part of the VDJ_GEX_matrix and combine
PlatypusDB_fetch(PlatypusDB.links = c("Kuhn2021a//ALL")
,save.to.disk = FALSE,load.to.enviroment = TRUE, load.to.list = FALSE
, combine.object = TRUE,path.to.save = "/Downloads")
#> Error in PlatypusDB_fetch(PlatypusDB.links = c("Kuhn2021a//ALL"), save.to.disk = FALSE,     load.to.enviroment = TRUE, load.to.list = FALSE, combine.object = TRUE,     path.to.save = "/Downloads"): One or more selected project ids were not present on the PlatypusDB. Please revise project id inputs

#Load VDJ dataframe of the VDJ GEX matrix for all samples of one project
loaded_list <- PlatypusDB_fetch(PlatypusDB.links = c("Kuhn2021a//VDJmatrix")
,save.to.disk = FALSE,load.to.enviroment = FALSE, load.to.list = TRUE)
#> Error in PlatypusDB_fetch(PlatypusDB.links = c("Kuhn2021a//VDJmatrix"),     save.to.disk = FALSE, load.to.enviroment = FALSE, load.to.list = TRUE): One or more selected project ids were not present on the PlatypusDB. Please revise project id inputs

#Load the VDJ and GEX RData of 2 samples from
#2 different projects which can be directly passed
#on to the VDJ_GEX_matrix function to integrate
#downloaded_objects <- PlatypusDB_fetch(
#PlatypusDB.links = c("Project1/s1/ALL", "Project1/s2/ALL")
#,save.to.disk = FALSE,load.to.enviroment = FALSE, load.to.list = TRUE
#, combine.objects = TRUE)

#integrated_samples <- VDJ_GEX_matrix_DB(data.in = downloaded_objects)

#Download metadata objects for projects
list_of_metadata_tables <- PlatypusDB_fetch(
PlatypusDB.links = c("Kuhn2021a//metadata")
,save.to.disk = FALSE,load.to.enviroment = FALSE, load.to.list = TRUE)
#> Error in PlatypusDB_fetch(PlatypusDB.links = c("Kuhn2021a//metadata"),     save.to.disk = FALSE, load.to.enviroment = FALSE, load.to.list = TRUE): One or more selected project ids were not present on the PlatypusDB. Please revise project id inputs

#Dowload of airr_rearrangement.tsv
#Load VDJ.RData into a list
#downloaded_objects <- PlatypusDB_fetch(
#PlatypusDB.links = c("Project1/ALL/VDJ.RData"),save.to.disk = FALSE
#,load.to.enviroment = FALSE, load.to.list = TRUE)

#Extract airr_rearrangement table for sample 1
#airr_rearrangement <- downloaded_objects[[1]][[1]][[6]]
#Index hierarchy: Sample, VDJ or GEX, VDJ element

#Save for import to AIRR compatible pipeline
#write.table(airr_rearrangement, file = "airr_rearrangement_s1.tsv", sep='\t',
#row.names = FALSE, quote=FALSE)

# }